birds dataset

The Swiss breeding bird survey ("Monitoring Häufige Brutvögel" MHB) has monitored the populations of 158 common species since 1999. The MHB sample consists of 267 1-km squares that are laid out as a grid across Switzerland. Fieldwork is conducted by about 200 skilled birdwatchers, most of them volunteers. Avian populations are monitored using a simplified territory mapping protocol, where each square is surveyed up to three times during the breeding season (only twice above the tree line). Surveys are conducted along a transect that does not change over the years.

The birds dataset has the data for 2014, except one quadrat not surveyed in 2014.

data("birds")

Format

A data frame with 266 observations on the following 166 variables.

158 bird species named in latin and whose occurrences are indicated as the number of visits to each site during which the species was observed, including 13 species not recorded in the year 2014 :

english name	latin name

Little Grebe	Tachybaptus ruficollis

Great Crested Grebe	Podiceps cristatus

Grey Heron	Ardea cinerea

Little Bittern	Ixobrychus minutus

White Stork	Ciconia ciconia

Mute Swan	Cygnus olor

Greylag Goose	Anser anser

Mallard	Anas platyrhynchos

Common Merganser	Mergus merganser

European Honey Buzzard	Pernis apivorus

Red Kite	Milvus milvus

Black Kite	Milvus migrans

Northern Goshawk	Accipiter gentilis

Eurasian Sparrowhawk	Accipiter nisus

Common Buzzard	Buteo buteo

Golden Eagle	Aquila chrysaetos

Peregrine Falcon	Falco peregrinus

Eurasian Hobby	Falco subbuteo

Common Kestrel	Falco tinnunculus

Western Capercaillie	Tetrao urogallus

Black Grouse	Tetrao tetrix

Rock Ptarmigan	Lagopus muta

Hazel Grouse	Bonasa bonasia

Rock Partridge	Alectoris graeca

Common Quail	Coturnix coturnix

Common Pheasant	Phasianus colchicus

Water Rail	Rallus aquaticus

Common Moorhen	Gallinula chloropus

Eurasian Coot	Fulica atra

Northern Lapwing	Vanellus vanellus

Little Ringed Plover	Charadrius dubius

Common Sandpiper	Actitis hypoleucos

Eurasian Woodcock	Scolopax rusticola

Yellow-legged Gull	Larus michahellis

Feral Pigeon	Columba livia domestica

Stock Dove	Columba oenas

Common Wood Pigeon	Columba palumbus

European Turtle Dove	Streptopelia turtur

Eurasian Collared Dove	Streptopelia decaocto

Common Cuckoo	Cuculus canorus

Barn Owl	Tyto alba

Scops Owl	Otus scops

Eurasian Eagle-Owl	Bubo bubo

Eurasian Pygmy Owl	Glaucidium passerinum

Little Owl	Athene noctua

Tawny Owl	Strix aluco

Long-eared Owl	Asio otus

Boreal Owl	Aegolius funereus

Alpine Swift	Apus melba

Common Swift	Apus apus

Common Kingfisher	Alcedo atthis

Eurasian Hoopoe	Upupa epops

Eurasian Wryneck	Jynx torquilla

European Green Woodpecker	Picus viridis

Grey-headed Woodpecker	Picus canus

Black Woodpecker	Dryocopus martius

Great Spotted Woodpecker	Dendrocopos major

Middle Spotted Woodpecker	Dendrocopos medius

White-backed Woodpecker	Dendrocopos leucotos

Lesser Spotted Woodpecker	Dendrocopos minor

Eurasian Three-toed Woodpecker	Picoides tridactylus

Woodlark	Lullula arborea

Eurasian Skylark	Alauda arvensis

Barn Swallow	Hirundo rustica

Eurasian Crag Martin	Ptyonoprogne rupestris

Common House Martin	Delichon urbicum

Eurasian Golden Oriole	Oriolus oriolus

Northern Raven	Corvus corax

Carrion/Hooded Crow	Corvus corone

Rook	Corvus frugilegus

Western Jackdaw	Corvus monedula

Eurasian Magpie	Pica pica

Spotted Nutcracker	Nucifraga caryocatactes

Eurasian Jay	Garrulus glandarius

Red-billed Chough	Pyrrhocorax pyrrhocorax

Alpine Chough	Pyrrhocorax graculus

Great Tit	Parus major

Blue Tit	Parus caeruleus

Coal Tit	Parus ater

European Crested Tit	Parus cristatus

Marsh Tit	Parus palustris

Willow Tit	Parus montanus

Long-tailed Bushtit	Aegithalos caudatus

Eurasian Nuthatch	Sitta europaea

Eurasian Treecreeper	Certhia familiaris

Short-toed Treecreeper	Certhia brachydactyla

Wallcreeper	Tichodroma muraria

White-throated Dipper	Cinclus cinclus

Winter Wren	Troglodytes troglodytes

European Robin	Erithacus rubecula

Common Nightingale	Luscinia megarhynchos

Bluethroat	Luscinia svecica

Black Redstart	Phoenicurus ochruros

Common Redstart	Phoenicurus phoenicurus

Whinchat	Saxicola rubetra

Eurasian Stonechat	Saxicola torquatus

Northern Wheatear	Oenanthe oenanthe

Rufous-tailed Rock Thrush	Monticola saxatilis

Ring Ouzel	Turdus torquatus

Common Blackbird	Turdus merula

Fieldfare	Turdus pilaris

Song Thrush	Turdus philomelos

Mistle Thrush	Turdus viscivorus

Common Grasshopper Warbler	Locustella naevia

Savi's Warbler	Locustella luscinioides

Great Reed Warbler	Acrocephalus arundinaceus

Eurasian Reed Warbler	Acrocephalus scirpaceus

Marsh Warbler	Acrocephalus palustris

Melodious Warbler	Hippolais polyglotta

Icterine Warbler	Hippolais icterina

Eurasian Blackcap	Sylvia atricapilla

Garden Warbler	Sylvia borin

Common Whitethroat	Sylvia communis

Lesser Whitethroat	Sylvia curruca

Willow Warbler	Phylloscopus trochilus

Common Chiffchaff	Phylloscopus collybita

Western Bonelli's Warbler	Phylloscopus bonelli

Wood Warbler	Phylloscopus sibilatrix

Goldcrest	Regulus regulus

Firecrest	Regulus ignicapilla

Spotted Flycatcher	Muscicapa striata

European Pied Flycatcher	Ficedula hypoleuca

Collared Flycatcher	Ficedula albicollis

Red-breasted Flycatcher	Ficedula parva

Dunnock	Prunella modularis

Alpine Accentor	Prunella collaris

Meadow Pipit	Anthus pratensis

Tawny Pipit	Anthus campestris

Tree Pipit	Anthus trivialis

Water Pipit	Anthus spinoletta

White Wagtail	Motacilla alba

Grey Wagtail	Motacilla cinerea

Western Yellow Wagtail	Motacilla flava

Woodchat Shrike	Lanius senator

Red-backed Shrike	Lanius collurio

Common Starling	Sturnus vulgaris

House Sparrow	Passer domesticus

Italian Sparrow	Passer hispaniolensis italiae

Eurasian Tree Sparrow	Passer montanus

White-winged Snowfinch	Montifringilla nivalis

Hawfinch	Coccothraustes coccothraustes

European Greenfinch	Carduelis chloris

European Goldfinch	Carduelis carduelis

Eurasian Siskin	Carduelis spinus

Common Linnet	Carduelis cannabina

Lesser Redpoll	Carduelis cabaret

Citril Finch	Serinus citrinella

European Serin	Serinus serinus

Eurasian Bullfinch	Pyrrhula pyrrhula

Common Rosefinch	Carpodacus erythrinus

Red Crossbill	Loxia curvirostra

Common Chaffinch	Fringilla coelebs

Corn Bunting	Emberiza calandra

Yellowhammer	Emberiza citrinella

Cirl Bunting	Emberiza cirlus

Ortolan Bunting	Emberiza hortulana

Rock Bunting	Emberiza cia

Common Reed Bunting	Emberiza schoeniclus

5 covariates collected on the 266 1x1 km quadrat as well as their identifiers and coordinates :

siteID	an alphanumeric site identifier

coordx	a numeric vector indicating the x coordinate of the centre of the quadrat; the coordinate reference system intentionally not specified.

coordy	a numeric vector indicating the y coordinate of the centre of the quadrat.

elev	a numeric vector indicating the mean elevation of the quadrat (m).

rlength	the length of the route walked in the quadrat (km).

nsurvey	a numeric vector indicating the number of replicate surveys planned in the quadrat; above the tree-line 2, otherwise 3.

forest	a numeric vector indicating the percentage of forest cover in the quadrat.

obs14	a categorical vector indicating the identifying number of the observer.

Source

Swiss Ornithological Institute

References

Kéry and Royle (2016) Applied Hierarachical Modeling in Ecology Section 11.3

Examples

data(birds, package="jSDM")
head(birds)

#>   Tachybaptus ruficollis Podiceps cristatus Ardea cinerea Ixobrychus minutus
#> 1                      0                  0             0                  0
#> 2                      0                  0             0                  0
#> 3                      0                  0             0                  0
#> 4                      0                  0             0                  0
#> 5                      0                  0             0                  0
#> 6                      0                  0             0                  0
#>   Ciconia ciconia Cygnus olor Anser anser Anas platyrhynchos Mergus merganser
#> 1               0           0           0                  0                0
#> 2               0           0           0                  0                0
#> 3               0           0           0                  3                0
#> 4               0           0           0                  0                0
#> 5               0           0           0                  0                0
#> 6               0           0           0                  0                0
#>   Pernis apivorus Milvus milvus Milvus migrans Accipiter gentilis
#> 1               0             0              3                  0
#> 2               0             0              3                  0
#> 3               1             1              1                  0
#> 4               0             0              2                  0
#> 5               0             2              2                  0
#> 6               0             0              3                  0
#>   Accipiter nisus Buteo buteo Aquila chrysaetos Falco peregrinus Falco subbuteo
#> 1               0           0                 0                0              0
#> 2               0           1                 0                0              0
#> 3               0           2                 0                0              0
#> 4               1           1                 0                0              0
#> 5               0           2                 0                0              0
#> 6               0           2                 0                0              1
#>   Falco tinnunculus Tetrao urogallus Tetrao tetrix Lagopus muta Bonasa bonasia
#> 1                 1                0             0            0              0
#> 2                 0                0             0            0              0
#> 3                 2                0             0            0              0
#> 4                 0                0             0            0              0
#> 5                 2                0             0            0              0
#> 6                 3                0             0            0              0
#>   Alectoris graeca Coturnix coturnix Phasianus colchicus Rallus aquaticus
#> 1                0                 0                   0                0
#> 2                0                 0                   0                0
#> 3                0                 0                   0                0
#> 4                0                 1                   0                0
#> 5                0                 0                   0                0
#> 6                0                 0                   0                0
#>   Gallinula chloropus Fulica atra Vanellus vanellus Charadrius dubius
#> 1                   0           0                 0                 0
#> 2                   0           0                 0                 0
#> 3                   0           2                 0                 0
#> 4                   0           0                 0                 0
#> 5                   0           0                 0                 0
#> 6                   0           0                 0                 0
#>   Actitis hypoleucos Scolopax rusticola Larus michahellis
#> 1                  0                  0                 0
#> 2                  0                  0                 0
#> 3                  0                  0                 0
#> 4                  0                  0                 0
#> 5                  0                  0                 0
#> 6                  0                  0                 0
#>   Columba livia domestica Columba oenas Columba palumbus Streptopelia turtur
#> 1                       3             1                3                   0
#> 2                       0             0                3                   0
#> 3                       0             1                3                   0
#> 4                       0             1                3                   0
#> 5                       0             0                3                   0
#> 6                       1             0                3                   0
#>   Streptopelia decaocto Cuculus canorus Tyto alba Otus scops Bubo bubo
#> 1                     3               0         0          0         0
#> 2                     0               1         0          0         0
#> 3                     0               1         0          0         0
#> 4                     0               0         0          0         0
#> 5                     0               0         0          0         0
#> 6                     2               0         0          0         0
#>   Glaucidium passerinum Athene noctua Strix aluco Asio otus Aegolius funereus
#> 1                     0             0           0         0                 0
#> 2                     0             0           0         0                 0
#> 3                     0             0           0         0                 0
#> 4                     0             0           0         0                 0
#> 5                     0             0           0         0                 0
#> 6                     0             0           0         0                 0
#>   Apus melba Apus apus Alcedo atthis Upupa epops Jynx torquilla Picus viridis
#> 1          0         2             0           0              0             0
#> 2          0         0             0           0              0             3
#> 3          0         0             0           0              0             0
#> 4          0         0             0           0              0             1
#> 5          0         0             0           0              0             3
#> 6          0         2             0           0              0             3
#>   Picus canus Dryocopus martius Dendrocopos major Dendrocopos medius
#> 1           0                 0                 3                  0
#> 2           0                 0                 3                  0
#> 3           0                 2                 1                  0
#> 4           0                 0                 2                  0
#> 5           0                 1                 1                  0
#> 6           0                 0                 0                  0
#>   Dendrocopos leucotos Dendrocopos minor Picoides tridactylus Lullula arborea
#> 1                    0                 0                    0               0
#> 2                    0                 0                    0               0
#> 3                    0                 0                    0               0
#> 4                    0                 0                    0               0
#> 5                    0                 0                    0               0
#> 6                    0                 0                    0               0
#>   Alauda arvensis Hirundo rustica Ptyonoprogne rupestris Delichon urbicum
#> 1               0               0                      0                2
#> 2               3               2                      0                2
#> 3               0               2                      0                2
#> 4               2               2                      0                1
#> 5               0               1                      0                1
#> 6               3               3                      0                1
#>   Oriolus oriolus Corvus corax Corvus corone Corvus frugilegus Corvus monedula
#> 1               0            0             3                 0               0
#> 2               0            0             3                 0               0
#> 3               0            0             3                 0               0
#> 4               0            0             3                 0               0
#> 5               0            0             3                 0               0
#> 6               0            0             3                 0               0
#>   Pica pica Nucifraga caryocatactes Garrulus glandarius Pyrrhocorax pyrrhocorax
#> 1         3                       0                   2                       0
#> 2         0                       0                   2                       0
#> 3         0                       0                   2                       0
#> 4         3                       0                   2                       0
#> 5         0                       0                   2                       0
#> 6         3                       0                   1                       0
#>   Pyrrhocorax graculus Parus major Parus caeruleus Parus ater Parus cristatus
#> 1                    0           3               3          0               0
#> 2                    0           3               3          0               0
#> 3                    0           3               2          3               3
#> 4                    0           3               3          3               1
#> 5                    0           3               2          3               3
#> 6                    0           3               3          0               0
#>   Parus palustris Parus montanus Aegithalos caudatus Sitta europaea
#> 1               0              0                   0              2
#> 2               0              0                   0              3
#> 3               3              3                   1              2
#> 4               0              0                   0              0
#> 5               1              1                   0              3
#> 6               1              0                   0              2
#>   Certhia familiaris Certhia brachydactyla Tichodroma muraria Cinclus cinclus
#> 1                  0                     3                  0               0
#> 2                  0                     3                  0               0
#> 3                  1                     0                  0               0
#> 4                  0                     1                  0               0
#> 5                  2                     0                  0               0
#> 6                  0                     2                  0               0
#>   Troglodytes troglodytes Erithacus rubecula Luscinia megarhynchos
#> 1                       0                  3                     0
#> 2                       3                  3                     0
#> 3                       3                  3                     0
#> 4                       3                  2                     0
#> 5                       3                  3                     0
#> 6                       0                  3                     0
#>   Luscinia svecica Phoenicurus ochruros Phoenicurus phoenicurus
#> 1                0                    3                       1
#> 2                0                    3                       0
#> 3                0                    3                       0
#> 4                0                    3                       1
#> 5                0                    3                       0
#> 6                0                    3                       1
#>   Saxicola rubetra Saxicola torquatus Oenanthe oenanthe Monticola saxatilis
#> 1                0                  0                 0                   0
#> 2                0                  0                 0                   0
#> 3                1                  0                 0                   0
#> 4                0                  0                 0                   0
#> 5                0                  0                 0                   0
#> 6                0                  3                 0                   0
#>   Turdus torquatus Turdus merula Turdus pilaris Turdus philomelos
#> 1                0             3              0                 0
#> 2                0             3              0                 3
#> 3                1             3              3                 3
#> 4                0             3              0                 1
#> 5                2             3              3                 3
#> 6                0             3              0                 0
#>   Turdus viscivorus Locustella naevia Locustella luscinioides
#> 1                 0                 0                       0
#> 2                 0                 0                       0
#> 3                 3                 0                       0
#> 4                 2                 0                       0
#> 5                 3                 0                       0
#> 6                 0                 0                       0
#>   Acrocephalus arundinaceus Acrocephalus scirpaceus Acrocephalus palustris
#> 1                         0                       0                      0
#> 2                         0                       0                      0
#> 3                         0                       0                      2
#> 4                         0                       1                      0
#> 5                         0                       0                      0
#> 6                         0                       0                      0
#>   Hippolais polyglotta Hippolais icterina Sylvia atricapilla Sylvia borin
#> 1                    0                  0                  3            0
#> 2                    0                  0                  3            0
#> 3                    0                  0                  3            0
#> 4                    0                  0                  3            2
#> 5                    0                  0                  2            1
#> 6                    0                  0                  3            1
#>   Sylvia communis Sylvia curruca Phylloscopus trochilus Phylloscopus collybita
#> 1               0              0                      0                      0
#> 2               0              0                      0                      3
#> 3               0              0                      3                      3
#> 4               0              0                      0                      2
#> 5               0              0                      1                      2
#> 6               3              0                      0                      0
#>   Phylloscopus bonelli Phylloscopus sibilatrix Regulus regulus
#> 1                    0                       0               0
#> 2                    0                       0               0
#> 3                    0                       1               3
#> 4                    0                       0               1
#> 5                    0                       1               3
#> 6                    0                       0               0
#>   Regulus ignicapilla Muscicapa striata Ficedula hypoleuca Ficedula albicollis
#> 1                   1                 1                  0                   0
#> 2                   3                 1                  1                   0
#> 3                   3                 0                  0                   0
#> 4                   1                 0                  0                   0
#> 5                   3                 0                  0                   0
#> 6                   0                 1                  0                   0
#>   Ficedula parva Prunella modularis Prunella collaris Anthus pratensis
#> 1              0                  0                 0                0
#> 2              0                  0                 0                0
#> 3              0                  2                 0                0
#> 4              0                  0                 0                0
#> 5              0                  2                 0                0
#> 6              0                  0                 0                0
#>   Anthus campestris Anthus trivialis Anthus spinoletta Motacilla alba
#> 1                 0                0                 0              3
#> 2                 0                0                 0              3
#> 3                 0                3                 0              3
#> 4                 0                0                 0              0
#> 5                 0                0                 0              3
#> 6                 0                0                 0              1
#>   Motacilla cinerea Motacilla flava Lanius senator Lanius collurio
#> 1                 0               0              0               0
#> 2                 0               0              0               0
#> 3                 0               0              0               1
#> 4                 0               0              0               0
#> 5                 0               0              0               1
#> 6                 0               0              0               3
#>   Sturnus vulgaris Passer domesticus Passer hispaniolensis italiae
#> 1                3                 3                             0
#> 2                3                 3                             0
#> 3                0                 3                             0
#> 4                3                 3                             0
#> 5                3                 1                             0
#> 6                3                 3                             0
#>   Passer montanus Montifringilla nivalis Coccothraustes coccothraustes
#> 1               0                      0                             0
#> 2               0                      0                             0
#> 3               0                      0                             0
#> 4               3                      0                             0
#> 5               0                      0                             0
#> 6               2                      0                             0
#>   Carduelis chloris Carduelis carduelis Carduelis spinus Carduelis cannabina
#> 1                 3                   3                0                   0
#> 2                 3                   3                0                   0
#> 3                 1                   0                0                   0
#> 4                 2                   1                0                   0
#> 5                 2                   2                0                   0
#> 6                 2                   3                0                   3
#>   Carduelis cabaret Serinus citrinella Serinus serinus Pyrrhula pyrrhula
#> 1                 0                  0               3                 0
#> 2                 0                  0               2                 0
#> 3                 0                  0               0                 3
#> 4                 0                  0               2                 0
#> 5                 0                  0               1                 0
#> 6                 0                  0               3                 0
#>   Carpodacus erythrinus Loxia curvirostra Fringilla coelebs Emberiza calandra
#> 1                     0                 0                 2                 0
#> 2                     0                 0                 3                 0
#> 3                     0                 1                 3                 0
#> 4                     0                 0                 3                 0
#> 5                     0                 0                 3                 0
#> 6                     0                 0                 3                 1
#>   Emberiza citrinella Emberiza cirlus Emberiza hortulana Emberiza cia
#> 1                   0               0                  0            0
#> 2                   0               0                  0            0
#> 3                   0               0                  0            0
#> 4                   3               0                  0            0
#> 5                   0               0                  0            0
#> 6                   3               0                  0            0
#>   Emberiza schoeniclus siteID coordx coordy elev rlength nsurvey forest obs14
#> 1                    0   Q001 922942  63276  450     6.4       3      3   386
#> 2                    0   Q002 928942  79276  450     5.5       3     21   147
#> 3                    2   Q003 928942 103276 1050     4.3       3     32    77
#> 4                    0   Q004 934942  95276  950     4.5       3      9   293
#> 5                    0   Q005 934942 111276 1150     5.4       3     35    77
#> 6                    0   Q006 946942  95276  550     3.6       3      2   361

# find species not recorded in 2014
which(colSums(birds[,1:158])==0)

#>        Ixobrychus minutus         Charadrius dubius                 Tyto alba 
#>                         4                        31                        41 
#>                 Bubo bubo             Athene noctua               Upupa epops 
#>                        43                        45                        52 
#>      Dendrocopos leucotos          Luscinia svecica Acrocephalus arundinaceus 
#>                        59                        92                       106 
#>            Ficedula parva            Lanius senator     Carpodacus erythrinus 
#>                       124                       134                       150 
#>        Emberiza hortulana 
#>                       156