eucalypts dataset

The Eucalyptus data set includes 12 taxa recorded in 458 plots spanning elevation gradients in the Grampians National Park, Victoria, which is known for high species diversity and endemism. The park has three mountain ranges interspersed with alluvial valleys and sand sheet and has a semi-Mediterranean climate with warm, dry summers and cool, wet winters.

This dataset records presence or absence at 458 sites of 12 eucalypts species, 7 covariates collected at these sites as well as their longitude and latitude.

data("eucalypts")

Format

A data frame with 458 observations on the following 21 variables.

12 eucalypts species which presence on sites is indicated by a 1 and absence by a 0 :
ALA

a binary vector indicating the occurrence of the species E. alaticaulis

ARE

a binary vector indicating the occurrence of the species E. arenacea

BAX

a binary vector indicating the occurrence of the species E. baxteri

CAM

a binary vector indicating the occurrence of the species E. camaldulensis

GON

a binary vector indicating the occurrence of the species E. goniocalyx

MEL

a binary vector indicating the occurrence of the species E. melliodora

OBL

a binary vector indicating the occurrence of the species E. oblique

OVA

a binary vector indicating the occurrence of the species E. ovata

WIL

a binary vector indicating the occurrence of the species E. willisii subsp. Falciformis

ALP

a binary vector indicating the occurrence of the species E. serraensis, E. verrucata and E. victoriana

VIM

a binary vector indicating the occurrence of the species E. viminalis subsp. Viminalis and Cygnetensis

ARO.SAB

a binary vector indicating the occurrence of the species E. aromaphloia and E. sabulosa

7 covariates collected on the 458 sites and their coordinates :
Rockiness

a numeric vector taking values from 0 to 95 corresponding to the rock cover of the site in percent estimated in 5 % increments in field plots

Sandiness

a binary vector indicating if soil texture categorie is sandiness based on soil texture classes from field plots and according to relative amounts of sand, silt, and clay particles

VallyBotFlat

a numeric vector taking values from 0 to 6 corresponding to the valley bottom flatness GIS-derived variable defining flat areas relative to surroundings likely to accumulate sediment (units correspond to the percentage of slope e.g. 0.5 = 16 %slope, 4.5 = 1 %slope, 5.5 = 0.5 %slope)

PPTann

a numeric vector taking values from 555 to 1348 corresponding to annual precipitation in millimeters measured as the sum of monthly precipitation estimated using BIOCLIM based on 20m grid cell Digital Elevation Model

Loaminess

a binary vector indicating if soil texture categorie is loaminess based on soil texture classes from field plots and according to relative amounts of sand, silt, and clay particles

cvTemp

a numeric vector taking values from 136 to 158 corresponding to coefficient of variation of temperature seasonality in percent measured as the standard deviation of weekly mean temperatures as a percentage of the annual mean temperature from BIOCLIM

T0

a numeric vector corresponding to solar radiation in \(WH/m^2\) measured as the amount of incident solar energy based on the visible sky and the sun's position. Derived from Digital Elevation Model in ArcGIS 9.2 Spatial Analyst for the summer solstice (December 22)

latitude

a numeric vector indicating the latitude of the studied site

longitude

a numeric vector indicating the longitude of the studied site

Source

Wilkinson, D. P.; Golding, N.; Guillera-Arroita, G.; Tingley, R. and McCarthy, M. A. (2018) A comparison of joint species distribution models for presence-absence data. Methods in Ecology and Evolution.

Examples

data(eucalypts, package="jSDM")
head(eucalypts)
#>   ALA ARE BAX CAM GON MEL OBL OVA WIL ALP VIM ARO.SAB Rockiness Sandiness
#> 1   0   0   0   0   0   0   0   0   1   1   0       0        60         1
#> 2   0   0   0   0   0   0   1   0   1   1   0       0        75         1
#> 3   0   0   1   0   0   0   0   0   1   1   0       0        70         1
#> 4   0   0   1   0   0   0   0   0   1   0   0       0        40         1
#> 5   0   0   1   0   0   0   1   0   0   0   0       0        15         1
#> 6   0   0   0   0   0   0   0   0   1   1   0       0        80         1
#>   VallyBotFlat PPTann Loaminess cvTemp      T0 latitude longitude
#> 1            0    785         0    142 6124.01 -37.4269  142.4574
#> 2            0    785         0    142 6124.01 -37.4269  142.4574
#> 3            0    780         0    142 3252.96 -37.4263  142.4577
#> 4            0    778         0    142 1636.63 -37.4269  142.4574
#> 5            0    772         0    142 1352.08 -37.4259  142.4584
#> 6            0    841         0    142 5018.48 -37.4224  142.4590