joint species distribution models

jSDM is an R package for fitting joint species distribution models (jSDM) in a hierarchical Bayesian framework.

The Gibbs sampler is written in C++. It uses Rcpp, Armadillo and GSL to maximize computation efficiency.

Package:	jSDM

Type:	Package

Version:	0.1.0

Date:	2019-01-11

License:	GPL-3

LazyLoad:	yes

Details

The package includes the following functions to fit various species distribution models :

function	data-type

jSDM_binomial_logit	presence-absence

jSDM_binomial_probit_block	presence-absence

jSDM_poisson_log	abundance

jSDM_binomial_probit_block :

Ecological process: $$y_{ij} \sim \mathcal{B}ernoulli(\theta_{ij})$$ where

if n_latent=0 and site_effect="none"	probit\((\theta_{ij}) = \beta_{0j} + X_i \beta_j\)

if n_latent>0 and site_effect="none"	probit\((\theta_{ij}) = \beta_{0j} + X_i \beta_j + W_i \lambda_j\)

if n_latent=0 and site_effect="fixed"	probit\((\theta_{ij}) = \beta_{0j} + X_i \beta_j + \alpha_i\) and \(\alpha_i \sim \mathcal{N}(0,V_\alpha)\)

if n_latent>0 and site_effect="fixed"	probit\((\theta_{ij}) = \beta_{0j} + X_i \beta_j + W_i \lambda_j + \alpha_i\)

if n_latent=0 and site_effect="random"	probit\((\theta_{ij}) = \beta_{0j} + X_i \beta_j + \alpha_i\)

if n_latent>0 and site_effect="random"	probit\((\theta_{ij}) = \beta_{0j} + X_i \beta_j + W_i \lambda_j + \alpha_i\) and \(\alpha_i \sim \mathcal{N}(0,V_\alpha)\)

jSDM_binomial_logit :

Ecological process : $$y_{ij} \sim \mathcal{B}inomial(\theta_{ij},t_i)$$ where

if n_latent=0 and site_effect="none"	logit\((\theta_{ij}) = \beta_{0j} + X_i \beta_j\)

if n_latent>0 and site_effect="none"	logit\((\theta_{ij}) = \beta_{0j} + X_i \beta_j + W_i \lambda_j\)

if n_latent=0 and site_effect="fixed"	logit\((\theta_{ij}) = \beta_{0j} + X_i \beta_j + \alpha_i\)

if n_latent>0 and site_effect="fixed"	logit\((\theta_{ij}) = \beta_{0j} + X_i \beta_j + W_i \lambda_j + \alpha_i\)

if n_latent=0 and site_effect="random"	logit\((\theta_{ij}) = \beta_{0j} + X_i \beta_j + \alpha_i\) and \(\alpha_i \sim \mathcal{N}(0,V_\alpha)\)

if n_latent>0 and site_effect="random"	logit\((\theta_{ij}) = \beta_{0j} + X_i \beta_j + W_i \lambda_j + \alpha_i\) and \(\alpha_i \sim \mathcal{N}(0,V_\alpha)\)

jSDM_poisson_log :

Ecological process : $$y_{ij} \sim \mathcal{P}oisson(\theta_{ij})$$ where

if n_latent=0 and site_effect="none"	log\((\theta_{ij}) = \beta_{0j} + X_i \beta_j\)

if n_latent>0 and site_effect="none"	log\((\theta_{ij}) = \beta_{0j} + X_i \beta_j + W_i \lambda_j\)

if n_latent=0 and site_effect="fixed"	log\((\theta_{ij}) = \beta_{0j} + X_i \beta_j + \alpha_i\)

if n_latent>0 and site_effect="fixed"	log\((\theta_{ij}) = \beta_{0j} + X_i \beta_j + W_i \lambda_j + \alpha_i\)

if n_latent=0 and site_effect="random"	log\((\theta_{ij}) = \beta_{0j} + X_i \beta_j + \alpha_i\) and \(\alpha_i \sim \mathcal{N}(0,V_\alpha)\)

if n_latent>0 and site_effect="random"	log\((\theta_{ij}) = \beta_{0j} + X_i \beta_j + W_i \lambda_j + \alpha_i\) and \(\alpha_i \sim \mathcal{N}(0,V_\alpha)\)

#'

jSDM_binomial_probit_block_long_format :

Ecological process: $$y_n \sim \mathcal{B}ernoulli(\theta_n)$$ such as \(species_n=j\) and \(site_n=i\), where

if n_latent=0 and site_effect="none"	probit\((\theta_{n}) = D_n \beta + X_n \beta_{sp.j}\)

if n_latent>0 and site_effect="none"	probit\((\theta_{n}) = D_n \beta + X_n \beta_{sp.j} + W_i \lambda_j\)

if n_latent=0 and site_effect="fixed"	probit\((\theta_{n}) = D_n \beta + X_n \beta_{sp.j} + \alpha_i\) and \(\alpha_i \sim \mathcal{N}(0,V_\alpha)\)

if n_latent>0 and site_effect="fixed"	probit\((\theta_{n}) = D_n \beta + X_n \beta_{sp.j} + W_i \lambda_j + \alpha_i\)

if n_latent=0 and site_effect="random"	probit\((\theta_{n}) = D_n \beta + X_n \beta_{sp.j} + \alpha_i\)

if n_latent>0 and site_effect="random"	probit\((\theta_{n}) = D_n \beta + X_n \beta_{sp.j} + W_i \lambda_j + \alpha_i\) and \(\alpha_i \sim \mathcal{N}(0,V_\alpha)\)

Author

Ghislain Vieilledent <ghislain.vieilledent@cirad.fr>

Jeanne Clément <jeanne.clement16@laposte.net>