Binomial logistic regression joint species distribution model

The jSDM_binomial_logit function performs a Binomial logistic regression in a Bayesian framework. The function calls a Gibbs sampler written in C++ code which uses an adaptive Metropolis algorithm to estimate the conditional posterior distribution of model's parameters.

jSDM_binomial_logit(
  burnin = 5000,
  mcmc = 10000,
  thin = 5,
  presence_site_sp,
  site_suitability,
  site_data,
  trials,
  beta_start = 0,
  mu_beta = 0,
  V_beta = 1e+06,
  ropt = 0.44,
  seed = 1234,
  verbose = 1
)

Arguments

burnin
The number of burnin iterations for the sampler.

mcmc
The number of Gibbs iterations for the sampler. Total number of Gibbs iterations is equal to burnin+mcmc. burnin+mcmc must be divisible by 10 and superior or equal to 100 so that the progress bar can be displayed.

thin
The thinning interval used in the simulation. The number of mcmc iterations must be divisible by this value.

presence_site_sp
A vector indicating the number of successes (or presences) and the absence by a zero for each species at studied sites.

site_suitability
A one-sided formula of the form '~x1+...+xp' with p terms specifying the explicative variables for the suitability process of the model.

site_data
data frame containing the model's explicative variables.

trials
A vector indicating the number of trials for each site. \(t_i\) should be superior or equal to \(y_ij\), the number of successes for observation \(n\). If \(t_i =0\), then \(y_ij=0\).

beta_start
Starting values for beta parameters of the suitability process. If beta_start takes a scalar value, then that value will serve for all of the betas.

mu_beta
Means of the priors for the \(\beta\) parameters of the suitability process. mubeta must be either a scalar or a p-length vector. If mubeta takes a scalar value, then that value will serve as the prior mean for all of the betas. The default value is set for an uninformative prior.

V_beta
Variances of the Normal priors for the \(\beta\) parameters of the suitability process. Vbeta must be either a scalar or a p-length vector. If Vbeta takes a scalar value, then that value will serve as the prior variance for all of the betas. The default variance is large and set to 1.0E6 for an uninformative flat prior.

ropt
Target acceptance rate for the adaptive Metropolis algorithm. Default to 0.44.

seed
The seed for the random number generator. Default to 1234.

verbose
A switch (0,1) which determines whether or not the progress of the sampler is printed to the screen. Default is 1: a progress bar is printed, indicating the step (in %) reached by the Gibbs sampler.

Value

An object of class "jSDM" acting like a list including :

mcmc.betas	An mcmc object that contains the posterior sample of estimated species effects. This object can be summarized by functions provided by the coda package.

mcmc.Deviance	The posterior sample of the deviance \(D\), with \(D=-2\log(\prod_ij P(y_ij|\beta_j,t_i))\), is also provided.

theta_latent	Predictive posterior mean of the probability associated to the suitability process for each observation.

model_spec	Various attributes of the model fitted, including the response and model matrix used, distributional assumptions as link function and family, trial sizes, hyperparameters used in the Bayesian estimation and mcmc, burnin and thin.

Details

We model an ecological process where the presence or absence of species \(j\) on site \(i\) is explained by habitat suitability.

Ecological process: $$y_ij \sim \mathcal{B}inomial(\theta_ij,t_i)$$ $$logit(\theta_ij) = X_i \beta_j$$

References

Gelfand, A. E.; Schmidt, A. M.; Wu, S.; Silander, J. A.; Latimer, A. and Rebelo, A. G. (2005) Modelling species diversity through species level hierarchical modelling. Applied Statistics, 54, 1-20.

Latimer, A. M.; Wu, S. S.; Gelfand, A. E. and Silander, J. A. (2006) Building statistical models to analyze species distributions. Ecological Applications, 16, 33-50.

See also

Examples

#==============================================
# jSDM_binomial_logit()
# Example with simulated data
#==============================================

#=================
#== Load libraries
library(jSDM)

#==================
#== Data simulation

#= Number of sites
nsite <- 100
#= Number of species
nsp <- 50
#= Set seed for repeatability
seed <- 1234

#= Number of visits associated to each site
set.seed(seed)
visits <- rpois(nsite,3)
visits[visits==0] <- 1

#= Ecological process (suitability)
x1 <- rnorm(nsite,0,1)
set.seed(2*seed)
x2 <- rnorm(nsite,0,1)
X <- cbind(rep(1,nsite),x1,x2)
np <- ncol(X)
beta.target <- matrix(runif(nsp*np,-2,2), byrow=TRUE, nrow=nsp)
logit.theta <- X %*% t(beta.target)
theta <- inv_logit(logit.theta)
set.seed(seed)
Y <- apply(theta, 2, rbinom, n=nsite, size=visits)

#= Site-occupancy model

mod <- jSDM_binomial_logit(# Chains
                           burnin=100,
                           mcmc=100,
                           thin=1,
                           # Response variable 
                           presence_site_sp=Y,
                           trials=visits,
                           # Explanatory variables
                           site_suitability=~x1+x2,
                           site_data=X,
                           # Starting values 
                           beta_start=0,
                           # Priors 
                           mu_beta=0, V_beta=1.0E6,
                           # Various
                           seed=1234, ropt=0.44, verbose=1)

#> 
#> Running the Gibbs sampler. It may be long, please keep cool :)
#> 
#> **********:10.0%, mean accept. rates= beta:0.262 
#> **********:20.0%, mean accept. rates= beta:0.271 
#> **********:30.0%, mean accept. rates= beta:0.365 
#> **********:40.0%, mean accept. rates= beta:0.419 
#> **********:50.0%, mean accept. rates= beta:0.436 
#> **********:60.0%, mean accept. rates= beta:0.441 
#> **********:70.0%, mean accept. rates= beta:0.438 
#> **********:80.0%, mean accept. rates= beta:0.441 
#> **********:90.0%, mean accept. rates= beta:0.429 
#> **********:100.0%, mean accept. rates= beta:0.424 

#== Outputs
# Parameter estimates
summary(mod$mcmc)

#> Length  Class   Mode 
#>      0   NULL   NULL 

pdf(file=file.path(tempdir(), "Posteriors_jSDM_binomial.pdf"))
plot(mod$mcmc.Deviance)
for(j in 1:nsp) plot(mod$mcmc.betas[[j]])
dev.off()

#> pdf 
#>   2 

#= Predictions
summary(mod$theta_latent)

#>        V1                V2                V3               V4        
#>  Min.   :0.01189   Min.   :0.02749   Min.   :0.1096   Min.   :0.1735  
#>  1st Qu.:0.16804   1st Qu.:0.24906   1st Qu.:0.7383   1st Qu.:0.6437  
#>  Median :0.46653   Median :0.43692   Median :0.8943   Median :0.8188  
#>  Mean   :0.46486   Mean   :0.45370   Mean   :0.8248   Mean   :0.7477  
#>  3rd Qu.:0.71815   3rd Qu.:0.67208   3rd Qu.:0.9594   3rd Qu.:0.9129  
#>  Max.   :0.98646   Max.   :0.96342   Max.   :0.9981   Max.   :0.9907  
#>        V5                 V6                V7                V8         
#>  Min.   :0.007143   Min.   :0.06378   Min.   :0.05565   Min.   :0.02258  
#>  1st Qu.:0.085340   1st Qu.:0.17524   1st Qu.:0.35258   1st Qu.:0.32288  
#>  Median :0.182008   Median :0.26033   Median :0.63202   Median :0.70753  
#>  Mean   :0.237880   Mean   :0.28566   Mean   :0.57927   Mean   :0.61435  
#>  3rd Qu.:0.368990   3rd Qu.:0.36642   3rd Qu.:0.81640   3rd Qu.:0.89322  
#>  Max.   :0.786053   Max.   :0.59430   Max.   :0.97563   Max.   :0.99642  
#>        V9              V10                V11              V12         
#>  Min.   :0.5916   Min.   :0.006826   Min.   :0.2897   Min.   :0.01741  
#>  1st Qu.:0.6411   1st Qu.:0.051886   1st Qu.:0.6014   1st Qu.:0.10272  
#>  Median :0.6553   Median :0.127312   Median :0.6992   Median :0.17735  
#>  Mean   :0.6561   Mean   :0.217602   Mean   :0.6777   Mean   :0.24354  
#>  3rd Qu.:0.6724   3rd Qu.:0.318720   3rd Qu.:0.7809   3rd Qu.:0.33579  
#>  Max.   :0.7084   Max.   :0.841786   Max.   :0.9037   Max.   :0.77955  
#>       V13                V14              V15              V16         
#>  Min.   :0.005176   Min.   :0.3596   Min.   :0.2009   Min.   :0.03823  
#>  1st Qu.:0.075956   1st Qu.:0.7722   1st Qu.:0.5886   1st Qu.:0.30232  
#>  Median :0.196991   Median :0.8793   Median :0.7759   Median :0.54471  
#>  Mean   :0.298754   Mean   :0.8342   Mean   :0.7191   Mean   :0.51158  
#>  3rd Qu.:0.488745   3rd Qu.:0.9358   3rd Qu.:0.8667   3rd Qu.:0.72298  
#>  Max.   :0.928115   Max.   :0.9811   Max.   :0.9792   Max.   :0.97276  
#>       V17              V18              V19                V20         
#>  Min.   :0.1140   Min.   :0.1245   Min.   :0.006405   Min.   :0.05687  
#>  1st Qu.:0.4921   1st Qu.:0.4575   1st Qu.:0.266173   1st Qu.:0.52997  
#>  Median :0.6651   Median :0.5955   Median :0.579495   Median :0.75055  
#>  Mean   :0.6256   Mean   :0.5946   Mean   :0.566148   Mean   :0.67468  
#>  3rd Qu.:0.7743   3rd Qu.:0.7687   3rd Qu.:0.888830   3rd Qu.:0.89240  
#>  Max.   :0.9243   Max.   :0.9522   Max.   :0.995116   Max.   :0.98030  
#>       V21              V22               V23              V24         
#>  Min.   :0.4624   Min.   :0.03499   Min.   :0.2006   Min.   :0.02809  
#>  1st Qu.:0.6420   1st Qu.:0.62484   1st Qu.:0.7539   1st Qu.:0.26245  
#>  Median :0.6973   Median :0.84852   Median :0.8851   Median :0.57457  
#>  Mean   :0.6872   Mean   :0.76098   Mean   :0.8248   Mean   :0.56573  
#>  3rd Qu.:0.7427   3rd Qu.:0.95756   3rd Qu.:0.9459   3rd Qu.:0.83649  
#>  Max.   :0.8250   Max.   :0.99878   Max.   :0.9955   Max.   :0.99641  
#>       V25                V26                V27               V28        
#>  Min.   :0.003868   Min.   :0.001485   Min.   :0.03491   Min.   :0.3031  
#>  1st Qu.:0.063579   1st Qu.:0.047743   1st Qu.:0.47109   1st Qu.:0.4934  
#>  Median :0.206370   Median :0.135325   Median :0.74244   Median :0.5756  
#>  Mean   :0.252622   Mean   :0.259797   Mean   :0.67507   Mean   :0.5794  
#>  3rd Qu.:0.400434   3rd Qu.:0.437626   3rd Qu.:0.90575   3rd Qu.:0.6696  
#>  Max.   :0.949173   Max.   :0.948166   Max.   :0.99462   Max.   :0.8104  
#>       V29              V30               V31              V32         
#>  Min.   :0.3801   Min.   :0.05452   Min.   :0.0312   Min.   :0.01368  
#>  1st Qu.:0.4201   1st Qu.:0.39083   1st Qu.:0.1262   1st Qu.:0.19822  
#>  Median :0.4330   Median :0.63861   Median :0.2380   Median :0.37180  
#>  Mean   :0.4356   Mean   :0.58799   Mean   :0.2791   Mean   :0.43096  
#>  3rd Qu.:0.4493   3rd Qu.:0.80640   3rd Qu.:0.3902   3rd Qu.:0.65383  
#>  Max.   :0.4903   Max.   :0.98298   Max.   :0.7979   Max.   :0.96817  
#>       V33               V34               V35                V36        
#>  Min.   :0.03278   Min.   :0.01056   Min.   :0.005001   Min.   :0.2039  
#>  1st Qu.:0.13744   1st Qu.:0.11082   1st Qu.:0.082877   1st Qu.:0.5398  
#>  Median :0.22203   Median :0.27015   Median :0.209350   Median :0.6714  
#>  Mean   :0.24335   Mean   :0.33939   Mean   :0.290996   Mean   :0.6559  
#>  3rd Qu.:0.33630   3rd Qu.:0.51896   3rd Qu.:0.444668   3rd Qu.:0.8039  
#>  Max.   :0.71866   Max.   :0.96299   Max.   :0.910645   Max.   :0.9395  
#>       V37              V38               V39               V40          
#>  Min.   :0.3891   Min.   :0.08557   Min.   :0.01459   Min.   :0.005639  
#>  1st Qu.:0.6445   1st Qu.:0.61068   1st Qu.:0.10751   1st Qu.:0.093956  
#>  Median :0.7417   Median :0.83852   Median :0.18350   Median :0.228584  
#>  Mean   :0.7223   Mean   :0.74698   Mean   :0.24707   Mean   :0.284651  
#>  3rd Qu.:0.8290   3rd Qu.:0.94057   3rd Qu.:0.36065   3rd Qu.:0.436654  
#>  Max.   :0.9546   Max.   :0.99026   Max.   :0.76709   Max.   :0.870387  
#>       V41               V42               V43               V44         
#>  Min.   :0.07902   Min.   :0.03726   Min.   :0.05471   Min.   :0.02966  
#>  1st Qu.:0.75053   1st Qu.:0.27446   1st Qu.:0.36421   1st Qu.:0.36543  
#>  Median :0.91386   Median :0.53099   Median :0.75833   Median :0.62285  
#>  Mean   :0.80195   Mean   :0.50661   Mean   :0.64839   Mean   :0.58156  
#>  3rd Qu.:0.97671   3rd Qu.:0.74141   3rd Qu.:0.93472   3rd Qu.:0.81111  
#>  Max.   :0.99913   Max.   :0.95676   Max.   :0.99821   Max.   :0.98873  
#>       V45              V46              V47               V48         
#>  Min.   :0.2473   Min.   :0.2886   Min.   :0.08397   Min.   :0.05871  
#>  1st Qu.:0.3154   1st Qu.:0.6635   1st Qu.:0.43736   1st Qu.:0.37928  
#>  Median :0.3500   Median :0.7688   Median :0.62131   Median :0.56854  
#>  Mean   :0.3477   Mean   :0.7454   Mean   :0.59091   Mean   :0.56224  
#>  3rd Qu.:0.3755   3rd Qu.:0.8709   3rd Qu.:0.76039   3rd Qu.:0.78005  
#>  Max.   :0.4609   Max.   :0.9653   Max.   :0.93297   Max.   :0.95929  
#>       V49                V50         
#>  Min.   :0.009912   Min.   :0.01176  
#>  1st Qu.:0.169306   1st Qu.:0.09296  
#>  Median :0.357989   Median :0.20631  
#>  Mean   :0.393204   Mean   :0.27290  
#>  3rd Qu.:0.619273   3rd Qu.:0.37610  
#>  Max.   :0.933579   Max.   :0.86295  

pdf(file=file.path(tempdir(), "Pred-Init.pdf"))
plot(theta, mod$theta_latent,
    main="theta",xlab="obs", ylab="fitted")
abline(a=0 ,b=1, col="red")
dev.off()

#> pdf 
#>   2