Madagascar's forest inventory

Dataset compiled from the national forest inventories carried out on 753 sites on the island of Madagascar, listing the presence or absence of 555 plant species on each of these sites between 1994 and 1996. We use these forest inventories to calculate a matrix indicating the presence by a 1 and the absence by a 0 of the species at each site by removing observations for which the species is not identified. This presence-absence matrix therefore records the occurrences of 483 species at 751 sites.

Format

madagascar is a data frame with 751 rows corresponding to the inventory sites and 483 columns corresponding to the species whose presence or absence has been recorded on the sites.

sp_

1 to 483 indicate by a 0 the absence of the species on one site and by a 1 its presence

site

"1" to "753" inventory sites identifiers.

Examples

data(madagascar, package="jSDM")
head(madagascar)
#>   sp_1 sp_2 sp_3 sp_4 sp_5 sp_6 sp_7 sp_8 sp_9 sp_10 sp_11 sp_12 sp_13 sp_14
#> 1    0    0    0    0    0    0    0    0    0     0     0     0     0     0
#> 2    0    0    0    0    0    1    0    0    0     0     0     0     0     0
#> 3    0    0    0    0    0    0    0    0    1     0     0     0     0     0
#> 4    0    0    0    0    0    0    0    0    0     0     0     0     0     0
#> 5    0    0    0    0    0    0    0    0    0     0     0     0     0     0
#> 6    0    0    0    1    0    0    0    0    0     0     0     0     0     1
#>   sp_15 sp_16 sp_17 sp_18 sp_19 sp_20 sp_21 sp_22 sp_23 sp_24 sp_25 sp_26 sp_27
#> 1     0     0     0     0     0     0     1     0     0     0     0     0     0
#> 2     0     0     0     0     0     0     0     1     0     0     0     0     0
#> 3     0     0     0     0     0     0     0     1     0     0     0     0     0
#> 4     0     0     0     0     0     0     1     0     0     0     0     0     0
#> 5     0     0     0     0     0     0     0     1     0     0     0     0     0
#> 6     0     0     0     0     0     0     1     1     0     0     0     0     0
#>   sp_28 sp_29 sp_30 sp_31 sp_32 sp_33 sp_34 sp_35 sp_36 sp_37 sp_38 sp_39 sp_40
#> 1     0     0     0     0     0     0     0     0     0     0     0     0     0
#> 2     0     0     0     0     0     0     0     0     0     0     0     0     0
#> 3     0     0     0     0     0     0     0     0     0     0     0     0     0
#> 4     0     0     0     0     0     0     0     0     0     0     0     0     0
#> 5     0     0     0     0     0     0     0     0     0     0     0     0     0
#> 6     0     0     0     0     0     0     0     0     0     0     0     0     0
#>   sp_41 sp_42 sp_43 sp_44 sp_45 sp_46 sp_47 sp_48 sp_49 sp_50 sp_51 sp_52 sp_53
#> 1     0     0     0     1     0     0     0     0     0     0     0     0     0
#> 2     0     0     0     0     0     0     0     1     0     0     0     0     0
#> 3     0     0     0     0     0     0     0     0     0     0     0     0     0
#> 4     0     0     0     0     0     0     0     0     0     0     0     0     0
#> 5     0     0     0     0     0     0     0     0     0     0     0     0     0
#> 6     0     1     0     0     0     0     0     0     0     0     0     0     0
#>   sp_54 sp_55 sp_56 sp_57 sp_58 sp_59 sp_60 sp_61 sp_62 sp_63 sp_64 sp_65 sp_66
#> 1     0     0     0     0     0     0     0     0     0     0     0     0     0
#> 2     0     0     0     0     0     0     0     0     0     0     0     0     0
#> 3     0     0     0     0     0     0     0     0     0     0     0     0     0
#> 4     1     0     0     0     0     0     0     0     0     0     0     0     0
#> 5     0     0     0     0     0     0     0     0     0     0     1     0     0
#> 6     0     0     0     0     0     0     0     0     0     0     0     0     0
#>   sp_67 sp_68 sp_69 sp_70 sp_71 sp_72 sp_73 sp_74 sp_75 sp_76 sp_77 sp_78 sp_79
#> 1     0     0     0     0     0     0     0     0     0     0     0     0     0
#> 2     0     0     0     0     0     0     0     0     0     0     0     0     0
#> 3     0     0     0     0     0     0     0     0     0     0     0     0     0
#> 4     0     0     0     0     0     0     0     0     0     0     0     1     0
#> 5     0     0     0     0     0     0     0     0     0     0     0     0     0
#> 6     0     0     0     0     0     0     0     0     0     0     0     0     0
#>   sp_80 sp_81 sp_82 sp_83 sp_84 sp_85 sp_86 sp_87 sp_88 sp_89 sp_90 sp_91 sp_92
#> 1     0     0     0     0     1     0     0     0     0     0     0     1     0
#> 2     0     0     0     0     0     0     0     0     0     0     0     0     0
#> 3     0     0     0     0     0     0     0     0     0     0     0     0     0
#> 4     0     0     0     0     0     0     0     0     0     0     0     0     0
#> 5     0     0     0     0     0     0     0     0     0     0     0     0     0
#> 6     0     0     0     0     0     0     0     0     0     0     0     0     0
#>   sp_93 sp_94 sp_95 sp_96 sp_97 sp_98 sp_99 sp_100 sp_101 sp_102 sp_103 sp_104
#> 1     0     0     0     0     0     0     0      0      0      0      0      0
#> 2     0     0     0     0     0     0     0      0      0      0      0      0
#> 3     0     0     1     0     0     0     0      0      0      0      0      0
#> 4     0     0     0     0     0     0     0      0      0      0      0      0
#> 5     0     0     0     0     0     0     0      0      0      0      0      0
#> 6     0     0     1     0     0     0     0      0      0      0      0      0
#>   sp_105 sp_106 sp_107 sp_108 sp_109 sp_110 sp_111 sp_112 sp_113 sp_114 sp_115
#> 1      0      0      0      0      0      0      0      0      0      0      0
#> 2      0      0      0      0      0      1      0      0      0      0      0
#> 3      0      0      0      0      0      0      0      0      0      0      0
#> 4      0      0      0      0      0      0      0      0      0      0      0
#> 5      0      0      0      0      0      0      0      0      0      0      0
#> 6      0      0      0      0      0      0      0      0      0      0      0
#>   sp_116 sp_117 sp_118 sp_119 sp_120 sp_121 sp_122 sp_123 sp_124 sp_125 sp_126
#> 1      0      0      0      0      0      0      0      0      0      0      0
#> 2      0      0      0      0      0      0      0      0      0      0      0
#> 3      0      0      0      0      0      0      0      0      0      0      0
#> 4      0      0      0      0      0      0      0      0      0      0      0
#> 5      0      0      0      0      0      0      0      0      0      0      0
#> 6      0      0      0      0      0      0      0      0      0      0      0
#>   sp_127 sp_128 sp_129 sp_130 sp_131 sp_132 sp_133 sp_134 sp_135 sp_136 sp_137
#> 1      0      0      0      0      0      0      0      0      0      0      0
#> 2      0      0      0      0      0      0      0      0      0      0      0
#> 3      0      0      0      0      0      0      0      0      0      0      0
#> 4      0      0      0      0      0      0      0      0      0      0      0
#> 5      0      0      0      0      0      0      0      0      0      0      0
#> 6      0      0      1      0      0      0      0      0      0      0      0
#>   sp_138 sp_139 sp_140 sp_141 sp_142 sp_143 sp_144 sp_145 sp_146 sp_147 sp_148
#> 1      0      0      0      0      0      0      0      0      1      0      1
#> 2      0      0      0      0      0      0      0      0      0      1      1
#> 3      0      0      0      0      0      0      0      0      0      0      0
#> 4      0      0      0      0      0      0      0      0      1      1      1
#> 5      0      0      0      0      0      0      0      0      1      0      1
#> 6      0      0      0      0      0      0      0      0      0      0      0
#>   sp_149 sp_150 sp_151 sp_152 sp_153 sp_154 sp_155 sp_156 sp_157 sp_158 sp_159
#> 1      0      0      0      0      0      1      1      1      0      0      0
#> 2      0      0      1      0      0      0      0      0      0      0      1
#> 3      1      1      1      1      1      0      0      0      0      0      0
#> 4      0      0      0      0      0      0      0      0      0      0      0
#> 5      0      1      0      0      0      0      0      0      1      1      0
#> 6      0      0      0      0      0      0      0      0      0      0      0
#>   sp_160 sp_161 sp_162 sp_163 sp_164 sp_165 sp_166 sp_167 sp_168 sp_169 sp_170
#> 1      0      0      0      0      0      0      0      0      0      0      0
#> 2      1      0      0      0      0      0      0      0      0      0      0
#> 3      0      0      0      0      0      0      0      0      0      0      0
#> 4      0      0      0      0      0      0      0      0      0      0      0
#> 5      0      0      0      0      0      0      0      0      0      0      0
#> 6      0      1      0      1      0      0      0      0      0      0      0
#>   sp_171 sp_172 sp_173 sp_174 sp_175 sp_176 sp_177 sp_178 sp_179 sp_180 sp_181
#> 1      0      0      0      0      0      0      0      0      0      0      0
#> 2      0      0      0      0      0      0      0      0      0      0      0
#> 3      0      0      0      0      0      0      0      0      0      0      0
#> 4      0      0      0      0      0      0      0      0      0      0      0
#> 5      0      0      0      0      0      0      0      0      0      0      0
#> 6      0      0      0      0      0      0      0      0      0      0      0
#>   sp_182 sp_183 sp_184 sp_185 sp_186 sp_187 sp_188 sp_189 sp_190 sp_191 sp_192
#> 1      0      0      0      0      0      0      0      0      0      0      0
#> 2      0      0      0      0      0      0      0      0      0      0      0
#> 3      0      0      0      0      0      0      0      0      0      0      0
#> 4      0      0      0      0      0      0      0      0      0      0      0
#> 5      0      0      0      0      0      0      0      0      0      0      0
#> 6      0      0      0      0      0      0      0      0      0      0      0
#>   sp_193 sp_194 sp_195 sp_196 sp_197 sp_198 sp_199 sp_200 sp_201 sp_202 sp_203
#> 1      0      0      0      0      0      0      0      0      0      0      0
#> 2      0      0      0      0      0      0      0      0      0      0      0
#> 3      0      0      0      0      0      0      0      0      0      0      0
#> 4      0      0      0      0      0      0      0      0      0      0      0
#> 5      0      0      0      0      0      0      0      0      0      0      0
#> 6      0      0      0      0      0      0      0      0      0      0      0
#>   sp_204 sp_205 sp_206 sp_207 sp_208 sp_209 sp_210 sp_211 sp_212 sp_213 sp_214
#> 1      0      0      0      0      0      0      0      0      0      0      0
#> 2      0      0      0      0      0      0      0      0      0      0      0
#> 3      0      0      0      0      0      0      0      0      0      0      0
#> 4      0      0      0      0      0      0      0      0      0      0      0
#> 5      0      0      0      0      0      0      0      0      0      0      0
#> 6      0      0      0      0      0      0      0      0      0      0      0
#>   sp_215 sp_216 sp_217 sp_218 sp_219 sp_220 sp_221 sp_222 sp_223 sp_224 sp_225
#> 1      0      0      0      0      0      0      0      0      0      0      0
#> 2      0      0      0      0      0      0      0      0      0      0      0
#> 3      0      0      0      0      0      0      0      0      0      0      0
#> 4      0      0      0      0      0      0      0      0      0      0      0
#> 5      0      0      0      0      0      0      0      0      0      0      0
#> 6      0      0      0      0      0      0      0      0      0      0      0
#>   sp_226 sp_227 sp_228 sp_229 sp_230 sp_231 sp_232 sp_233 sp_234 sp_235 sp_236
#> 1      0      0      0      0      0      0      0      0      0      0      0
#> 2      0      0      0      0      0      0      0      0      0      0      0
#> 3      0      0      0      0      0      0      0      0      0      0      0
#> 4      0      0      0      0      0      0      0      0      0      0      0
#> 5      0      0      0      0      0      0      0      0      0      0      0
#> 6      0      0      0      0      0      0      0      0      0      0      0
#>   sp_237 sp_238 sp_239 sp_240 sp_241 sp_242 sp_243 sp_244 sp_245 sp_246 sp_247
#> 1      0      0      0      0      0      0      0      0      0      0      0
#> 2      0      0      0      0      0      0      0      0      0      0      0
#> 3      0      0      0      0      0      0      0      0      0      0      0
#> 4      0      0      0      0      0      0      0      0      0      0      0
#> 5      0      0      0      0      0      0      0      0      0      0      0
#> 6      0      0      0      0      0      0      0      0      0      0      0
#>   sp_248 sp_249 sp_250 sp_251 sp_252 sp_253 sp_254 sp_255 sp_256 sp_257 sp_258
#> 1      0      0      0      0      0      0      0      0      0      0      0
#> 2      0      0      0      0      0      0      0      0      0      0      0
#> 3      0      0      0      0      0      0      0      0      0      0      0
#> 4      0      0      0      0      0      0      0      0      0      0      0
#> 5      0      0      0      0      0      0      0      0      0      0      0
#> 6      0      0      0      0      0      0      0      0      0      0      0
#>   sp_259 sp_260 sp_261 sp_262 sp_263 sp_264 sp_265 sp_266 sp_267 sp_268 sp_269
#> 1      0      0      0      0      0      0      0      0      0      0      0
#> 2      0      0      0      0      0      0      0      0      0      0      0
#> 3      0      0      0      0      0      0      0      0      0      0      0
#> 4      0      0      0      0      0      0      0      0      0      0      0
#> 5      0      0      0      0      0      0      0      0      0      0      0
#> 6      0      0      0      0      0      0      0      0      0      0      0
#>   sp_270 sp_271 sp_272 sp_273 sp_274 sp_275 sp_276 sp_277 sp_278 sp_279 sp_280
#> 1      0      0      0      0      0      0      0      0      0      0      0
#> 2      0      0      0      0      0      0      0      0      0      0      0
#> 3      0      0      0      0      0      0      0      0      0      0      0
#> 4      0      0      0      0      0      0      0      0      0      0      0
#> 5      0      0      0      0      0      0      0      0      0      0      0
#> 6      0      0      0      0      0      0      0      0      0      0      0
#>   sp_281 sp_282 sp_283 sp_284 sp_285 sp_286 sp_287 sp_288 sp_289 sp_290 sp_291
#> 1      0      0      0      0      0      0      0      0      0      0      0
#> 2      0      0      0      0      0      0      0      0      0      0      0
#> 3      0      0      0      0      0      0      0      0      0      0      0
#> 4      0      0      0      0      0      0      0      0      0      0      0
#> 5      0      0      0      0      0      0      0      0      0      0      0
#> 6      0      0      0      0      0      0      0      0      0      0      0
#>   sp_292 sp_293 sp_294 sp_295 sp_296 sp_297 sp_298 sp_299 sp_300 sp_301 sp_302
#> 1      0      0      0      0      0      0      0      0      0      0      0
#> 2      0      0      0      0      0      0      0      0      0      0      0
#> 3      0      0      0      0      0      0      0      0      0      0      0
#> 4      0      0      0      0      0      0      0      0      0      0      0
#> 5      0      0      0      0      0      0      0      0      0      0      0
#> 6      0      0      0      0      0      0      0      0      0      0      0
#>   sp_303 sp_304 sp_305 sp_306 sp_307 sp_308 sp_309 sp_310 sp_311 sp_312 sp_313
#> 1      0      0      0      0      0      0      0      0      0      0      0
#> 2      0      0      0      0      0      0      0      0      0      0      0
#> 3      0      0      0      0      0      0      0      0      0      0      0
#> 4      0      0      0      0      0      0      0      0      0      0      0
#> 5      0      0      0      0      0      0      0      0      0      0      0
#> 6      0      0      0      0      0      0      0      0      0      0      0
#>   sp_314 sp_315 sp_316 sp_317 sp_318 sp_319 sp_320 sp_321 sp_322 sp_323 sp_324
#> 1      0      0      0      0      0      0      0      0      0      0      0
#> 2      0      0      0      0      0      0      0      0      0      0      0
#> 3      0      0      0      0      0      0      0      0      0      0      0
#> 4      0      0      0      0      0      0      0      0      0      0      0
#> 5      0      0      0      0      0      0      0      0      0      0      0
#> 6      0      0      0      0      0      0      0      0      0      0      0
#>   sp_325 sp_326 sp_327 sp_328 sp_329 sp_330 sp_331 sp_332 sp_333 sp_334 sp_335
#> 1      0      0      0      0      0      0      0      0      0      0      0
#> 2      0      0      0      0      0      0      0      0      0      0      0
#> 3      0      0      0      0      0      0      0      0      0      0      0
#> 4      0      0      0      0      0      0      0      0      0      0      0
#> 5      0      0      0      0      0      0      0      0      0      0      0
#> 6      0      0      0      0      0      0      0      0      0      0      0
#>   sp_336 sp_337 sp_338 sp_339 sp_340 sp_341 sp_342 sp_343 sp_344 sp_345 sp_346
#> 1      0      0      0      0      0      0      0      0      0      0      0
#> 2      0      0      0      0      0      0      0      0      0      0      0
#> 3      0      0      0      0      0      0      0      0      0      0      0
#> 4      0      0      0      0      0      0      0      0      0      0      0
#> 5      0      0      0      0      0      0      0      0      0      0      0
#> 6      0      0      0      0      0      0      0      0      0      0      0
#>   sp_347 sp_348 sp_349 sp_350 sp_351 sp_352 sp_353 sp_354 sp_355 sp_356 sp_357
#> 1      0      0      0      0      0      0      0      0      0      0      0
#> 2      0      0      0      0      0      0      0      0      0      0      0
#> 3      0      0      0      0      0      0      0      0      0      0      0
#> 4      0      0      0      0      0      0      0      0      0      0      0
#> 5      0      0      0      0      0      0      0      0      0      0      0
#> 6      0      0      0      0      0      0      0      0      0      0      0
#>   sp_358 sp_359 sp_360 sp_361 sp_362 sp_363 sp_364 sp_365 sp_366 sp_367 sp_368
#> 1      0      0      0      0      0      0      0      0      0      0      0
#> 2      0      0      0      0      0      0      0      0      0      0      0
#> 3      0      0      0      0      0      0      0      0      0      0      0
#> 4      0      0      0      0      0      0      0      0      0      0      0
#> 5      0      0      0      0      0      0      0      0      0      0      0
#> 6      0      0      0      0      0      0      0      0      0      0      0
#>   sp_369 sp_370 sp_371 sp_372 sp_373 sp_374 sp_375 sp_376 sp_377 sp_378 sp_379
#> 1      0      0      0      0      0      0      0      0      0      0      0
#> 2      0      0      0      0      0      0      0      0      0      0      0
#> 3      0      0      0      0      0      0      0      0      0      0      0
#> 4      0      0      0      0      0      0      0      0      0      0      0
#> 5      0      0      0      0      0      0      0      0      0      0      0
#> 6      0      0      0      0      0      0      0      0      0      0      0
#>   sp_380 sp_381 sp_382 sp_383 sp_384 sp_385 sp_386 sp_387 sp_388 sp_389 sp_390
#> 1      0      0      0      0      0      0      0      0      0      0      0
#> 2      0      0      0      0      0      0      0      0      0      0      0
#> 3      0      0      0      0      0      0      0      0      0      0      0
#> 4      0      0      0      0      0      0      0      0      0      0      0
#> 5      0      0      0      0      0      0      0      0      0      0      0
#> 6      0      0      0      0      0      0      0      0      0      0      0
#>   sp_391 sp_392 sp_393 sp_394 sp_395 sp_396 sp_397 sp_398 sp_399 sp_400 sp_401
#> 1      0      0      0      0      0      0      0      0      0      0      0
#> 2      0      0      0      0      0      0      0      0      0      0      0
#> 3      0      0      0      0      0      0      0      0      0      0      0
#> 4      0      0      0      0      0      0      0      0      0      0      0
#> 5      0      0      0      0      0      0      0      0      0      0      0
#> 6      0      0      0      0      0      0      0      0      0      0      0
#>   sp_402 sp_403 sp_404 sp_405 sp_406 sp_407 sp_408 sp_409 sp_410 sp_411 sp_412
#> 1      0      0      0      0      0      0      0      0      0      0      0
#> 2      0      0      0      0      0      0      0      0      0      0      0
#> 3      0      0      0      0      0      0      0      0      0      0      0
#> 4      0      0      0      0      0      0      0      0      0      0      0
#> 5      0      0      0      0      0      0      0      0      0      0      0
#> 6      0      0      0      0      0      0      0      0      0      0      0
#>   sp_413 sp_414 sp_415 sp_416 sp_417 sp_418 sp_419 sp_420 sp_421 sp_422 sp_423
#> 1      0      0      0      0      0      0      0      0      0      0      0
#> 2      0      0      0      0      0      0      0      0      0      0      0
#> 3      0      0      0      0      0      0      0      0      0      0      0
#> 4      0      0      0      0      0      0      0      0      0      0      0
#> 5      0      0      0      0      0      0      0      0      0      0      0
#> 6      0      0      0      0      0      0      0      0      0      0      0
#>   sp_424 sp_425 sp_426 sp_427 sp_428 sp_429 sp_430 sp_431 sp_432 sp_433 sp_434
#> 1      0      0      0      0      0      0      0      0      0      0      0
#> 2      0      0      0      0      0      0      0      0      0      0      0
#> 3      0      0      0      0      0      0      0      0      0      0      0
#> 4      0      0      0      0      0      0      0      0      0      0      0
#> 5      0      0      0      0      0      0      0      0      0      0      0
#> 6      0      0      0      0      0      0      0      0      0      0      0
#>   sp_435 sp_436 sp_437 sp_438 sp_439 sp_440 sp_441 sp_442 sp_443 sp_444 sp_445
#> 1      0      0      0      0      0      0      0      0      0      0      0
#> 2      0      0      0      0      0      0      0      0      0      0      0
#> 3      0      0      0      0      0      0      0      0      0      0      0
#> 4      0      0      0      0      0      0      0      0      0      0      0
#> 5      0      0      0      0      0      0      0      0      0      0      0
#> 6      0      0      0      0      0      0      0      0      0      0      0
#>   sp_446 sp_447 sp_448 sp_449 sp_450 sp_451 sp_452 sp_453 sp_454 sp_455 sp_456
#> 1      0      0      0      0      0      0      0      0      0      0      0
#> 2      0      0      0      0      0      0      0      0      0      0      0
#> 3      0      0      0      0      0      0      0      0      0      0      0
#> 4      0      0      0      0      0      0      0      0      0      0      0
#> 5      0      0      0      0      0      0      0      0      0      0      0
#> 6      0      0      0      0      0      0      0      0      0      0      0
#>   sp_457 sp_458 sp_459 sp_460 sp_461 sp_462 sp_463 sp_464 sp_465 sp_466 sp_467
#> 1      0      0      0      0      0      0      0      0      0      0      0
#> 2      0      0      0      0      0      0      0      0      0      0      0
#> 3      0      0      0      0      0      0      0      0      0      0      0
#> 4      0      0      0      0      0      0      0      0      0      0      0
#> 5      0      0      0      0      0      0      0      0      0      0      0
#> 6      0      0      0      0      0      0      0      0      0      0      0
#>   sp_468 sp_469 sp_470 sp_471 sp_472 sp_473 sp_474 sp_475 sp_476 sp_477 sp_478
#> 1      0      0      0      0      0      0      0      0      0      0      0
#> 2      0      0      0      0      0      0      0      0      0      0      0
#> 3      0      0      0      0      0      0      0      0      0      0      0
#> 4      0      0      0      0      0      0      0      0      0      0      0
#> 5      0      0      0      0      0      0      0      0      0      0      0
#> 6      0      0      0      0      0      0      0      0      0      0      0
#>   sp_479 sp_480 sp_481 sp_482 sp_483
#> 1      0      0      0      0      0
#> 2      0      0      0      0      0
#> 3      0      0      0      0      0
#> 4      0      0      0      0      0
#> 5      0      0      0      0      0
#> 6      0      0      0      0      0