mites dataset

This example data set is composed of 70 cores of mostly Sphagnum mosses collected on the territory of the Station de biologie des Laurentides of Université de Montréal, Québec, Canada in June 1989.

The whole sampling area was 2.5 m x 10 m in size and thirty-five taxa were recognized as species, though many were not given a species name, owing to the incomplete stage of systematic knowledge of the North American Oribatid fauna.

The data set comprises the abundances of 35 morphospecies, 5 substrate and micritopographic variables, and the x-y Cartesian coordinates of the 70 sampling sites.

See Borcard et al. (1992, 1994) for details.

data("mites")

Details

Oribatid mites (Acari: Oribatida) are a very diversified group of small (0.2–1.2 mm) soil-dwelling, mostly microphytophagous and detritivorous arthropods. A well aerated soil or a complex substrate like Sphagnum mosses present in bogs and wet forests can harbour up to several hundred thousand individuals per square metre. Local assemblages are sometimes composed of over a hundred species, including many rare ones. This diversity makes oribatid mites an interesting target group to study community–environment relationships at very local scales.

Format

A data frame with 70 observations on the following 42 variables.

Abundance of 35 Oribatid mites morphospecies named :

Brachy, PHTH, HPAV, RARD, SSTR, Protopl, MEGR, MPRO, TVIE, HMIN,HMIN2, NPRA, TVEL, ONOV, SUCT, LCIL, Oribatul1, Ceratoz1, PWIL, Galumna1, Steganacarus2, HRUF, Trhypochth1, PPEL, NCOR, SLAT, FSET, Lepidozetes, Eupelops, Minigalumna, LRUG, PLAG2, Ceratoz3, Oppia.minus and Trimalaco2.

5 covariates collected on the 70 sites and their coordinates :

substrate	a categorical vector indicating substrate type using a 7-level unordered factor : sph1, sph2, sph3, sph4, litter, peat and inter for interface.

shrubs	a categorical vector indicating shrub density using a 3-level ordered factor : None, Few and Many

topo	a categorical vector indicating microtopography using a factor with levels blanket and hummock

density	a numeric vector indicating the substrate density (g/L)

water	a numeric vector indicating the water content of the substrate (g/L)

x	a numeric vector indicating first coordinates of sampling sites

y	a numeric vector indicating second coordinates of sampling sites

Source

Pierre Legendre

References

Borcard, D.; Legendre, P. and Drapeau, P. (1992) Partialling out the spatial component of ecological variation. Ecology 73: 1045-1055.

Borcard, D. and Legendre, P. (1994) Environmental control and spatial structure in ecological communities: an example using Oribatid mites (Acari, Oribatei). Environmental and Ecological Statistics 1: 37-61.

Borcard, D. and Legendre, P. (2002) All-scale spatial analysis of ecological data by means of principal coordinates of neighbour matrices. Ecological Modelling 153: 51-68.

Examples

data(mites, package="jSDM")
head(mites)

#>   Brachy PHTH HPAV RARD SSTR Protopl MEGR MPRO TVIE HMIN HMIN2 NPRA TVEL ONOV
#> 1     17    5    5    3    2       1    4    2    2    1     4    1   17    4
#> 2      2    7   16    0    6       0    4    2    0    0     1    3   21   27
#> 3      4    3    1    1    2       0    3    0    0    0     6    3   20   17
#> 4     23    7   10    2    2       0    4    0    1    2    10    0   18   47
#> 5      5    8   13    9    0      13    0    0    0    3    14    3   32   43
#> 6     19    7    5    9    3       2    3    0    0   20    16    2   13   38
#>   SUCT LCIL Oribatul1 Ceratoz1 PWIL Galumna1 Steganacarus2 HRUF Trhypochth1
#> 1    9   50         3        1    1        8             0    0           0
#> 2   12  138         6        0    1        3             9    1           1
#> 3   10   89         3        0    2        1             8    0           3
#> 4   17  108        10        1    0        1             2    1           2
#> 5   27    5         1        0    5        2             1    0           1
#> 6   39    3         5        0    1        1             8    0           4
#>   PPEL NCOR SLAT FSET Lepidozetes Eupelops Minigalumna LRUG PLAG2 Ceratoz3
#> 1    0    0    0    0           0        0           0    0     0        0
#> 2    1    2    2    2           1        0           0    0     0        0
#> 3    0    2    0    8           0        0           0    0     0        0
#> 4    1    3    2   12           0        0           0    0     0        0
#> 5    0    0    0   12           2        0           0    0     0        0
#> 6    0    1    0   10           0        0           0    0     0        0
#>   Oppia.minus Trimalaco2 substrate shrubs    topo density   water   x   y
#> 1           0          0      sph1    few hummock  39.179 350.148 0.2 0.1
#> 2           0          0    litter    few hummock  54.989 434.815 1.0 0.1
#> 3           0          0     inter    few hummock  46.069 371.720 1.2 0.3
#> 4           0          0      sph1    few hummock  48.187 360.501 1.4 0.5
#> 5           0          0      sph1    few hummock  23.551 204.126 2.4 0.7
#> 6           0          0      sph1    few hummock  57.318 311.551 1.8 0.9