mosquitos dataset

Presence or absence at 167 sites of 16 species that constitute the aquatic faunal community studied, 13 covariates collected at these sites and their coordinates.

data("mosquitos")

Format

A data frame with 167 observations on the following 31 variables :

16 aquatic species including larvae of four mosquito species (all potential vectors of human disease), which presence on sites is indicated by a 1 and absence by a 0 :
Culex_pipiens_sl

a binary vector (mosquito species)

Culex_modestus

a binary vector (mosquito species)

Culiseta_annulata

a binary vector (mosquito species)

Anopheles_maculipennis_sl

a binary vector (mosquito species)

waterboatmen__Corixidae

a binary vector

diving_beetles__Dysticidae

a binary vector

damselflies__Zygoptera

a binary vector

swimming_beetles__Haliplidae

a binary vector

opossum_shrimps__Mysidae

a binary vector

ditch_shrimp__Gammarus

a binary vector

beetle_larvae__Coleoptera

a binary vector

dragonflies__Anisoptera

a binary vector

mayflies__Ephemeroptera

a binary vector

newts__Pleurodelinae

a binary vector

fish

a binary vector

saucer_bugs__Ilyocoris

a binary vector

13 covariates collected on the 167 sites and their coordinates :
depth__cm

a numeric vector corresponding to the water depth in cm recorded as the mean of the depth at the edge and the centre of each dip site

temperature__C

a numeric vector corresponding to the temperature in °C

oxidation_reduction_potential__Mv

a numeric vector corresponding to the redox potential of the water in millivolts (mV)

salinity__ppt

a numeric vector corresponding to the salinity of the water in parts per thousand (ppt)

High-resolution digital photographs were taken of vegetation at the edge and centre dip points and the presence or absence of different vegetation types at each dipsite was determined from these photographs using field guides :
water_crowfoot__Ranunculus

a binary vector indicating presence on sites by a 1 and absence by a 0 of the plant species Ranunculus aquatilis which common name is water-crowfoot

rushes__Juncus_or_Scirpus

a binary vector indicating presence on sites by a 1 and absence by a 0 of rushes from the Juncus or Scirpus genus

filamentous_algae

a binary vector indicating presence on sites by a 1 and absence by a 0 of filamentous algae

emergent_grass

a binary vector indicating presence on sites by a 1 and absence by a 0 of emergent grass

ivy_leafed_duckweed__Lemna_trisulca

a binary vector indicating presence on sites by a 1 and absence by a 0 of ivy leafed duckweed of species Lemna trisulca

bulrushes__Typha

a binary vector indicating presence on sites by a 1 and absence by a 0 of bulrushes from the Typha genus

reeds_Phragmites

a binary vector indicating presence on sites by a 1 and absence by a 0 of reeds from the Phragmites genus

marestail__Hippuris

a binary vector indicating presence on sites by a 1 and absence by a 0 of plants from the Hippuris genus known as mare's-tail

common_duckweed__Lemna_minor

a binary vector indicating presence on sites by a 1 and absence by a 0 of common duckweed of species Lemna minor

x

a numeric vector of first coordinates corresponding to each site

y

a numeric vector of second coordinates corresponding to each site

Source

Wilkinson, D. P.; Golding, N.; Guillera-Arroita, G.; Tingley, R. and McCarthy, M. A. (2018) A comparison of joint species distribution models for presence-absence data. Methods in Ecology and Evolution.

Examples

data(mosquitos, package="jSDM")
head(mosquitos)
#>   Culex_pipiens_sl Culex_modestus Culiseta_annulata Anopheles_maculipennis_sl
#> 1                0              0                 0                         1
#> 2                0              1                 0                         1
#> 3                1              1                 0                         1
#> 4                0              0                 0                         1
#> 5                1              1                 0                         1
#> 6                1              0                 0                         1
#>   waterboatmen__Corixidae diving_beetles__Dysticidae damselflies__Zygoptera
#> 1                       1                          0                      1
#> 2                       0                          1                      1
#> 3                       1                          0                      1
#> 4                       1                          0                      0
#> 5                       1                          0                      1
#> 6                       1                          0                      1
#>   swimming_beetles__Haliplidae opossum_shrimps__Mysidae ditch_shrimp__Gammarus
#> 1                            0                        0                      0
#> 2                            1                        0                      1
#> 3                            0                        0                      0
#> 4                            0                        0                      0
#> 5                            1                        0                      0
#> 6                            1                        0                      0
#>   beetle_larvae__Coleoptera dragonflies__Anisoptera mayflies__Ephemeroptera
#> 1                         1                       0                       1
#> 2                         0                       0                       0
#> 3                         0                       0                       1
#> 4                         0                       0                       0
#> 5                         1                       0                       0
#> 6                         0                       0                       1
#>   newts__Pleurodelinae fish saucer_bugs__Ilyocoris depth__cm temperature__C
#> 1                    0    0                      0  13.33333        22.5500
#> 2                    0    0                      0  34.37500        19.7500
#> 3                    1    0                      1  18.75000        18.3725
#> 4                    0    0                      0  36.87500        19.9475
#> 5                    0    0                      1  11.87500        21.0050
#> 6                    0    0                      0  17.50000        17.8925
#>   oxidation_reduction_potential__Mv salinity__ppt water_crowfoot__Ranunculus
#> 1                            -26.00     0.8333333                          0
#> 2                             -2.50     4.8900000                          1
#> 3                            -36.50     4.1750000                          1
#> 4                            -19.25     3.1025000                          0
#> 5                            -74.50     7.1050000                          0
#> 6                            -28.75     2.6250000                          0
#>   rushes__Juncus_or_Scirpus filamentous_algae emergent_grass
#> 1                         1                 1              1
#> 2                         1                 1              0
#> 3                         1                 1              0
#> 4                         1                 1              0
#> 5                         1                 1              0
#> 6                         1                 1              0
#>   ivy_leafed_duckweed__Lemna_trisulca bulrushes__Typha reeds_Phragmites
#> 1                                   0                0                0
#> 2                                   0                0                0
#> 3                                   0                0                0
#> 4                                   0                0                0
#> 5                                   0                0                1
#> 6                                   0                0                0
#>   marestail__Hippuris common_duckweed__Lemna_minor        x        y
#> 1                   1                            0 191.5561 59.87633
#> 2                   0                            0 191.5466 59.84163
#> 3                   0                            0 191.3421 59.84280
#> 4                   1                            0 191.2939 59.57974
#> 5                   0                            0 191.6692 59.79505
#> 6                   0                            1 191.3501 59.75232