Predict method for models fitted with jSDM

Predicted values for models fitted with jSDM

# S3 method for jSDM
predict(object, newdata=NULL, Id_species, Id_sites, type="mean",
  probs=c(0.025,0.975), ...)

Arguments

object
An object of class "jSDM".

newdata
An optional data frame in which to look for variables with which to predict. If omitted, the fitted values are used.

Id_species
An vector of character or integer indicating for which species the probabilities of presence on chosen sites will be predicted.

Id_sites
An vector of integer indicating for which sites the probabilities of presence of specified species will be predicted.

type
Type of prediction. Can be "mean" for predictive posterior mean, "quantile" for producing sample quantiles from the predictive posterior corresponding to the given probabilities (see probs argument) or "posterior" for the full predictive posterior for each prediction. Using "quantile" or "posterior" might lead to memory problem depending on the number of predictions and the number of samples for the jSDM model's parameters.

probs
Numeric vector of probabilities with values in [0,1] and used when type="quantile".

...
Further arguments passed to or from other methods.

Value

Return a vector for the predictive posterior mean when type="mean", a data-frame with the mean and quantiles when type="quantile" or an mcmc object (see coda package) with posterior distribution for each prediction when type="posterior".

See also