birds dataset

The Swiss breeding bird survey ("Monitoring Häufige Brutvögel" MHB) has monitored the populations of 158 common species since 1999. The MHB sample from data(MHB2014, package="AHMbook") consists of 267 1-km squares that are laid out as a grid across Switzerland. Fieldwork is conducted by about 200 skilled birdwatchers, most of them volunteers. Avian populations are monitored using a simplified territory mapping protocol, where each square is surveyed up to three times during the breeding season (only twice above the tree line). Surveys are conducted along a transect that does not change over the years.

The birds dataset has the data for 2014, except one quadrat not surveyed in 2014.

data("birds")

Format

A data frame with 266 observations on the following 166 variables.

158 bird species named in latin and whose occurrences are indicated as the number of visits to each site during which the species was observed, including 13 species not recorded in the year 2014 and 5 covariates collected on the 266 1x1 km quadrat as well as their identifiers and coordinates :

siteID	an alphanumeric site identifier

coordx	a numeric vector indicating the x coordinate of the centre of the quadrat; the coordinate reference system intentionally not specified.

coordy	a numeric vector indicating the y coordinate of the centre of the quadrat.

elev	a numeric vector indicating the mean elevation of the quadrat (m).

rlength	the length of the route walked in the quadrat (km).

nsurvey	a numeric vector indicating the number of replicate surveys planned in the quadrat; above the tree-line 2, otherwise 3.

forest	a numeric vector indicating the percentage of forest cover in the quadrat.

obs14	a categorical vector indicating the identifying number of the observer.

Source

Swiss Ornithological Institute

Details

Only the Latin names of bird species are given in this dataset but you can find the corresponding English names in the original dataset : codedata(MHB2014, package="AHMbook").

References

Kéry and Royle (2016) Applied Hierarachical Modeling in Ecology Section 11.3

Examples

data(birds, package="jSDM")
head(birds)
#>   Tachybaptus ruficollis Podiceps cristatus Ardea cinerea Ixobrychus minutus
#> 1                      0                  0             0                  0
#> 2                      0                  0             0                  0
#> 3                      0                  0             0                  0
#> 4                      0                  0             0                  0
#> 5                      0                  0             0                  0
#> 6                      0                  0             0                  0
#>   Ciconia ciconia Cygnus olor Anser anser Anas platyrhynchos Mergus merganser
#> 1               0           0           0                  0                0
#> 2               0           0           0                  0                0
#> 3               0           0           0                  3                0
#> 4               0           0           0                  0                0
#> 5               0           0           0                  0                0
#> 6               0           0           0                  0                0
#>   Pernis apivorus Milvus milvus Milvus migrans Accipiter gentilis
#> 1               0             0              3                  0
#> 2               0             0              3                  0
#> 3               1             1              1                  0
#> 4               0             0              2                  0
#> 5               0             2              2                  0
#> 6               0             0              3                  0
#>   Accipiter nisus Buteo buteo Aquila chrysaetos Falco peregrinus Falco subbuteo
#> 1               0           0                 0                0              0
#> 2               0           1                 0                0              0
#> 3               0           2                 0                0              0
#> 4               1           1                 0                0              0
#> 5               0           2                 0                0              0
#> 6               0           2                 0                0              1
#>   Falco tinnunculus Tetrao urogallus Tetrao tetrix Lagopus muta Bonasa bonasia
#> 1                 1                0             0            0              0
#> 2                 0                0             0            0              0
#> 3                 2                0             0            0              0
#> 4                 0                0             0            0              0
#> 5                 2                0             0            0              0
#> 6                 3                0             0            0              0
#>   Alectoris graeca Coturnix coturnix Phasianus colchicus Rallus aquaticus
#> 1                0                 0                   0                0
#> 2                0                 0                   0                0
#> 3                0                 0                   0                0
#> 4                0                 1                   0                0
#> 5                0                 0                   0                0
#> 6                0                 0                   0                0
#>   Gallinula chloropus Fulica atra Vanellus vanellus Charadrius dubius
#> 1                   0           0                 0                 0
#> 2                   0           0                 0                 0
#> 3                   0           2                 0                 0
#> 4                   0           0                 0                 0
#> 5                   0           0                 0                 0
#> 6                   0           0                 0                 0
#>   Actitis hypoleucos Scolopax rusticola Larus michahellis
#> 1                  0                  0                 0
#> 2                  0                  0                 0
#> 3                  0                  0                 0
#> 4                  0                  0                 0
#> 5                  0                  0                 0
#> 6                  0                  0                 0
#>   Columba livia domestica Columba oenas Columba palumbus Streptopelia turtur
#> 1                       3             1                3                   0
#> 2                       0             0                3                   0
#> 3                       0             1                3                   0
#> 4                       0             1                3                   0
#> 5                       0             0                3                   0
#> 6                       1             0                3                   0
#>   Streptopelia decaocto Cuculus canorus Tyto alba Otus scops Bubo bubo
#> 1                     3               0         0          0         0
#> 2                     0               1         0          0         0
#> 3                     0               1         0          0         0
#> 4                     0               0         0          0         0
#> 5                     0               0         0          0         0
#> 6                     2               0         0          0         0
#>   Glaucidium passerinum Athene noctua Strix aluco Asio otus Aegolius funereus
#> 1                     0             0           0         0                 0
#> 2                     0             0           0         0                 0
#> 3                     0             0           0         0                 0
#> 4                     0             0           0         0                 0
#> 5                     0             0           0         0                 0
#> 6                     0             0           0         0                 0
#>   Apus melba Apus apus Alcedo atthis Upupa epops Jynx torquilla Picus viridis
#> 1          0         2             0           0              0             0
#> 2          0         0             0           0              0             3
#> 3          0         0             0           0              0             0
#> 4          0         0             0           0              0             1
#> 5          0         0             0           0              0             3
#> 6          0         2             0           0              0             3
#>   Picus canus Dryocopus martius Dendrocopos major Dendrocopos medius
#> 1           0                 0                 3                  0
#> 2           0                 0                 3                  0
#> 3           0                 2                 1                  0
#> 4           0                 0                 2                  0
#> 5           0                 1                 1                  0
#> 6           0                 0                 0                  0
#>   Dendrocopos leucotos Dendrocopos minor Picoides tridactylus Lullula arborea
#> 1                    0                 0                    0               0
#> 2                    0                 0                    0               0
#> 3                    0                 0                    0               0
#> 4                    0                 0                    0               0
#> 5                    0                 0                    0               0
#> 6                    0                 0                    0               0
#>   Alauda arvensis Hirundo rustica Ptyonoprogne rupestris Delichon urbicum
#> 1               0               0                      0                2
#> 2               3               2                      0                2
#> 3               0               2                      0                2
#> 4               2               2                      0                1
#> 5               0               1                      0                1
#> 6               3               3                      0                1
#>   Oriolus oriolus Corvus corax Corvus corone Corvus frugilegus Corvus monedula
#> 1               0            0             3                 0               0
#> 2               0            0             3                 0               0
#> 3               0            0             3                 0               0
#> 4               0            0             3                 0               0
#> 5               0            0             3                 0               0
#> 6               0            0             3                 0               0
#>   Pica pica Nucifraga caryocatactes Garrulus glandarius Pyrrhocorax pyrrhocorax
#> 1         3                       0                   2                       0
#> 2         0                       0                   2                       0
#> 3         0                       0                   2                       0
#> 4         3                       0                   2                       0
#> 5         0                       0                   2                       0
#> 6         3                       0                   1                       0
#>   Pyrrhocorax graculus Parus major Parus caeruleus Parus ater Parus cristatus
#> 1                    0           3               3          0               0
#> 2                    0           3               3          0               0
#> 3                    0           3               2          3               3
#> 4                    0           3               3          3               1
#> 5                    0           3               2          3               3
#> 6                    0           3               3          0               0
#>   Parus palustris Parus montanus Aegithalos caudatus Sitta europaea
#> 1               0              0                   0              2
#> 2               0              0                   0              3
#> 3               3              3                   1              2
#> 4               0              0                   0              0
#> 5               1              1                   0              3
#> 6               1              0                   0              2
#>   Certhia familiaris Certhia brachydactyla Tichodroma muraria Cinclus cinclus
#> 1                  0                     3                  0               0
#> 2                  0                     3                  0               0
#> 3                  1                     0                  0               0
#> 4                  0                     1                  0               0
#> 5                  2                     0                  0               0
#> 6                  0                     2                  0               0
#>   Troglodytes troglodytes Erithacus rubecula Luscinia megarhynchos
#> 1                       0                  3                     0
#> 2                       3                  3                     0
#> 3                       3                  3                     0
#> 4                       3                  2                     0
#> 5                       3                  3                     0
#> 6                       0                  3                     0
#>   Luscinia svecica Phoenicurus ochruros Phoenicurus phoenicurus
#> 1                0                    3                       1
#> 2                0                    3                       0
#> 3                0                    3                       0
#> 4                0                    3                       1
#> 5                0                    3                       0
#> 6                0                    3                       1
#>   Saxicola rubetra Saxicola torquatus Oenanthe oenanthe Monticola saxatilis
#> 1                0                  0                 0                   0
#> 2                0                  0                 0                   0
#> 3                1                  0                 0                   0
#> 4                0                  0                 0                   0
#> 5                0                  0                 0                   0
#> 6                0                  3                 0                   0
#>   Turdus torquatus Turdus merula Turdus pilaris Turdus philomelos
#> 1                0             3              0                 0
#> 2                0             3              0                 3
#> 3                1             3              3                 3
#> 4                0             3              0                 1
#> 5                2             3              3                 3
#> 6                0             3              0                 0
#>   Turdus viscivorus Locustella naevia Locustella luscinioides
#> 1                 0                 0                       0
#> 2                 0                 0                       0
#> 3                 3                 0                       0
#> 4                 2                 0                       0
#> 5                 3                 0                       0
#> 6                 0                 0                       0
#>   Acrocephalus arundinaceus Acrocephalus scirpaceus Acrocephalus palustris
#> 1                         0                       0                      0
#> 2                         0                       0                      0
#> 3                         0                       0                      2
#> 4                         0                       1                      0
#> 5                         0                       0                      0
#> 6                         0                       0                      0
#>   Hippolais polyglotta Hippolais icterina Sylvia atricapilla Sylvia borin
#> 1                    0                  0                  3            0
#> 2                    0                  0                  3            0
#> 3                    0                  0                  3            0
#> 4                    0                  0                  3            2
#> 5                    0                  0                  2            1
#> 6                    0                  0                  3            1
#>   Sylvia communis Sylvia curruca Phylloscopus trochilus Phylloscopus collybita
#> 1               0              0                      0                      0
#> 2               0              0                      0                      3
#> 3               0              0                      3                      3
#> 4               0              0                      0                      2
#> 5               0              0                      1                      2
#> 6               3              0                      0                      0
#>   Phylloscopus bonelli Phylloscopus sibilatrix Regulus regulus
#> 1                    0                       0               0
#> 2                    0                       0               0
#> 3                    0                       1               3
#> 4                    0                       0               1
#> 5                    0                       1               3
#> 6                    0                       0               0
#>   Regulus ignicapilla Muscicapa striata Ficedula hypoleuca Ficedula albicollis
#> 1                   1                 1                  0                   0
#> 2                   3                 1                  1                   0
#> 3                   3                 0                  0                   0
#> 4                   1                 0                  0                   0
#> 5                   3                 0                  0                   0
#> 6                   0                 1                  0                   0
#>   Ficedula parva Prunella modularis Prunella collaris Anthus pratensis
#> 1              0                  0                 0                0
#> 2              0                  0                 0                0
#> 3              0                  2                 0                0
#> 4              0                  0                 0                0
#> 5              0                  2                 0                0
#> 6              0                  0                 0                0
#>   Anthus campestris Anthus trivialis Anthus spinoletta Motacilla alba
#> 1                 0                0                 0              3
#> 2                 0                0                 0              3
#> 3                 0                3                 0              3
#> 4                 0                0                 0              0
#> 5                 0                0                 0              3
#> 6                 0                0                 0              1
#>   Motacilla cinerea Motacilla flava Lanius senator Lanius collurio
#> 1                 0               0              0               0
#> 2                 0               0              0               0
#> 3                 0               0              0               1
#> 4                 0               0              0               0
#> 5                 0               0              0               1
#> 6                 0               0              0               3
#>   Sturnus vulgaris Passer domesticus Passer hispaniolensis italiae
#> 1                3                 3                             0
#> 2                3                 3                             0
#> 3                0                 3                             0
#> 4                3                 3                             0
#> 5                3                 1                             0
#> 6                3                 3                             0
#>   Passer montanus Montifringilla nivalis Coccothraustes coccothraustes
#> 1               0                      0                             0
#> 2               0                      0                             0
#> 3               0                      0                             0
#> 4               3                      0                             0
#> 5               0                      0                             0
#> 6               2                      0                             0
#>   Carduelis chloris Carduelis carduelis Carduelis spinus Carduelis cannabina
#> 1                 3                   3                0                   0
#> 2                 3                   3                0                   0
#> 3                 1                   0                0                   0
#> 4                 2                   1                0                   0
#> 5                 2                   2                0                   0
#> 6                 2                   3                0                   3
#>   Carduelis cabaret Serinus citrinella Serinus serinus Pyrrhula pyrrhula
#> 1                 0                  0               3                 0
#> 2                 0                  0               2                 0
#> 3                 0                  0               0                 3
#> 4                 0                  0               2                 0
#> 5                 0                  0               1                 0
#> 6                 0                  0               3                 0
#>   Carpodacus erythrinus Loxia curvirostra Fringilla coelebs Emberiza calandra
#> 1                     0                 0                 2                 0
#> 2                     0                 0                 3                 0
#> 3                     0                 1                 3                 0
#> 4                     0                 0                 3                 0
#> 5                     0                 0                 3                 0
#> 6                     0                 0                 3                 1
#>   Emberiza citrinella Emberiza cirlus Emberiza hortulana Emberiza cia
#> 1                   0               0                  0            0
#> 2                   0               0                  0            0
#> 3                   0               0                  0            0
#> 4                   3               0                  0            0
#> 5                   0               0                  0            0
#> 6                   3               0                  0            0
#>   Emberiza schoeniclus siteID coordx coordy elev rlength nsurvey forest obs14
#> 1                    0   Q001 922942  63276  450     6.4       3      3   386
#> 2                    0   Q002 928942  79276  450     5.5       3     21   147
#> 3                    2   Q003 928942 103276 1050     4.3       3     32    77
#> 4                    0   Q004 934942  95276  950     4.5       3      9   293
#> 5                    0   Q005 934942 111276 1150     5.4       3     35    77
#> 6                    0   Q006 946942  95276  550     3.6       3      2   361
# find species not recorded in 2014
which(colSums(birds[,1:158])==0)
#>        Ixobrychus minutus         Charadrius dubius                 Tyto alba 
#>                         4                        31                        41 
#>                 Bubo bubo             Athene noctua               Upupa epops 
#>                        43                        45                        52 
#>      Dendrocopos leucotos          Luscinia svecica Acrocephalus arundinaceus 
#>                        59                        92                       106 
#>            Ficedula parva            Lanius senator     Carpodacus erythrinus 
#>                       124                       134                       150 
#>        Emberiza hortulana 
#>                       156