fungi dataset

Presence or absence of 11 species of fungi on dead-wood objects at 800 sites and 12 covariates collected at these sites.

data("fungi")

Format

A data frame with 800 observations on the following 23 variables :

11 fungi species which presence on sites is indicated by a 1 and absence by a 0 :
antser

a binary vector

antsin

a binary vector

astfer

a binary vector

fompin

a binary vector

hetpar

a binary vector

junlut

a binary vector

phefer

a binary vector

phenig

a binary vector

phevit

a binary vector

poscae

a binary vector

triabi

a binary vector

12 covariates collected on the 800 sites :
diam

a numeric vector indicating the diameter of dead-wood object

dc1

a binary vector indicating if the decay class is 1 measured in the scale 1, 2, 3, 4, 5 (from freshly decayed to almost completely decayed)

dc2

a binary vector indicating if the decay class is 2

dc3

a binary vector indicating if the decay class is 3

dc4

a binary vector indicating if the decay class is 4

dc5

a binary vector indicating if the decay class is 5

quality3

a binary vector indicating if the quality is level 3

quality4

a binary vector indicating if the quality is level 4

ground3

a binary vector indicating if the ground contact is level 3 as 2 = no ground contact, 3 = less than half of the log in ground contact and 4 = more than half of the log in ground contact

ground4

a binary vector a binary vector indicating if the ground contact is level 4

epi

a numeric vector indicating the epiphyte cover

bark

a numeric vector indicating the bark cover

Source

Wilkinson, D. P.; Golding, N.; Guillera-Arroita, G.; Tingley, R. and McCarthy, M. A. (2018) A comparison of joint species distribution models for presence-absence data. Methods in Ecology and Evolution.

Examples

data(fungi, package="jSDM")
head(fungi)
#>   antser antsin astfer fompin hetpar junlut phefer phenig phevit poscae triabi
#> 1      0      0      0      0      0      0      1      0      0      0      0
#> 2      0      0      0      0      0      0      0      0      0      0      0
#> 3      0      0      0      0      0      0      0      0      1      0      0
#> 4      0      0      0      0      0      0      0      0      1      0      0
#> 5      0      0      0      0      0      0      0      0      0      0      0
#> 6      0      0      0      0      0      0      0      0      0      0      0
#>         diam dc1 dc2 dc3 dc4 dc5 quality3 quality4 ground3 ground4        epi
#> 1  1.3699426   1   0   0   0   0        0        1       0       1 -0.6626584
#> 2 -0.2083835   0   0   0   0   0        1        0       0       0 -0.6626584
#> 3  0.2435434   0   0   0   0   0        0        1       0       1 -0.3568567
#> 4  0.8040889   1   0   0   0   0        0        1       0       0 -0.6626584
#> 5 -0.5927520   0   0   0   0   0        0        1       0       0 -0.6626584
#> 6 -0.8227620   0   0   1   0   0        0        0       0       1 -0.3568567
#>         bark
#> 1  0.5297272
#> 2  1.2511949
#> 3 -1.0334527
#> 4 -0.9132081
#> 5  1.2511949
#> 6  1.2511949