Calculate the residual correlation matrix from a latent variable model (LVM)

This function use coefficients \((\lambda_{jl} with j=1,\dots,n_species and l=1,\dots,n_latent)\), corresponding to latent variables fitted using jSDM package, to calculate the variance-covariance matrix which controls correlation between species.

get_residual_cor(mod)

Arguments

mod
An object of class "jSDM"

Value

results A list including :

cov.mean	Average over the MCMC samples of the variance-covariance matrix

cov.median	Median over the MCMC samples of the variance-covariance matrix

cor.mean	Average over the MCMC samples of the residual correlation matrix

cor.median	Median over the MCMC samples of the residual correlation matrix

Details

After fitting the jSDM with latent variables, the fullspecies residual correlation matrix : \(R=(R_ij) avec i=1,\ldots, nspecies et j=1,\ldots, nspecies\) can be derived from the covariance in the latent variables such as :

\(\Sigma_{ij}\)	\(= \lambda_i .\lambda_j' + 1\)	if i=j

\(= \lambda_i .\lambda_j'\)	else,

then we compute correlations from covariances : $$R_{ij} = \frac{\Sigma_{ij}}{\sqrt{\Sigma_ii\Sigma _jj}}$$.

See also

Author

Ghislain Vieilledent <ghislain.vieilledent@cirad.fr>

Jeanne Clément <jeanne.clement16@laposte.net>

Examples

library(jSDM)
# frogs data
 data(frogs, package="jSDM")
 # Arranging data
 PA_frogs <- frogs[,4:12]
 # Normalized continuous variables
 Env_frogs <- cbind(scale(frogs[,1]),frogs[,2],scale(frogs[,3]))
 colnames(Env_frogs) <- colnames(frogs[,1:3])
 Env_frogs <- as.data.frame(Env_frogs)
 # Parameter inference
# Increase the number of iterations to reach MCMC convergence
mod <- jSDM_binomial_probit_block(# Response variable
                                  presence_site_sp=PA_frogs,
                                  # Explanatory variables
                                  site_suitability = ~.,
                                  site_data = Env_frogs,
                                  n_latent=2,
                                  site_effect="random",
                                  # Chains
                                  burnin=100,
                                  mcmc=100,
                                  thin=1,
                                  # Starting values
                                  alpha_start=0,
                                  beta_start=0,
                                  lambda_start=0,
                                  W_start=0,
                                  V_alpha=1,
                                  # Priors
                                  shape=0.5, rate=0.0005,
                                  mu_beta=0, V_beta=1.0E6,
                                  mu_lambda=0, V_lambda=10,
                                  # Various
                                  seed=1234, verbose=1)

#> 
#> Running the Gibbs sampler. It may be long, please keep cool :)
#> 
#> **********:10.0% 
#> **********:20.0% 
#> **********:30.0% 
#> **********:40.0% 
#> **********:50.0% 
#> **********:60.0% 
#> **********:70.0% 
#> **********:80.0% 
#> **********:90.0% 
#> **********:100.0% 

# Calcul of residual correlation between species 
 result <- get_residual_cor(mod)
 result$cov.mean

#>             [,1]       [,2]       [,3]       [,4]       [,5]      [,6]
#>  [1,]  2.3079891  0.7471740  1.6803794 -0.4037165  1.1899267  3.275969
#>  [2,]  0.7471740  1.8296132 -0.0652573 -0.5490315  0.8148569  2.327582
#>  [3,]  1.6803794 -0.0652573 16.8915634  1.5741158  1.0809211  3.282450
#>  [4,] -0.4037165 -0.5490315  1.5741158  1.7300967 -0.5229658 -1.412468
#>  [5,]  1.1899267  0.8148569  1.0809211 -0.5229658  2.3063015  3.408416
#>  [6,]  3.2759688  2.3275825  3.2824500 -1.4124677  3.4084165 10.747239
#>  [7,]  1.3615326  0.8181101  3.2624211 -0.2569041  1.3734775  3.944970
#>  [8,] -0.6059024 -0.8891704  1.6070053  0.8188570 -0.8102590 -2.334552
#>  [9,]  2.1639516  0.9451873  5.6079821 -0.0926125  1.9375572  5.479210
#>             [,7]       [,8]       [,9]
#>  [1,]  1.3615326 -0.6059024  2.1639516
#>  [2,]  0.8181101 -0.8891704  0.9451873
#>  [3,]  3.2624211  1.6070053  5.6079821
#>  [4,] -0.2569041  0.8188570 -0.0926125
#>  [5,]  1.3734775 -0.8102590  1.9375572
#>  [6,]  3.9449701 -2.3345522  5.4792103
#>  [7,]  3.1294777 -0.5921022  2.8210637
#>  [8,] -0.5921022  2.4809059 -0.4384975
#>  [9,]  2.8210637 -0.4384975  5.7418445

 result$cor.mean

#>             [,1]        [,2]        [,3]        [,4]       [,5]       [,6]
#>  [1,]  1.0000000  0.33277477  0.27850383 -0.18257340  0.4456666  0.6013028
#>  [2,]  0.3327748  1.00000000 -0.01437364 -0.26125254  0.3571910  0.4933452
#>  [3,]  0.2785038 -0.01437364  1.00000000  0.28850792  0.1795431  0.2670346
#>  [4,] -0.1825734 -0.26125254  0.28850792  1.00000000 -0.2332198 -0.3012252
#>  [5,]  0.4456666  0.35719097  0.17954310 -0.23321976  1.0000000  0.6133917
#>  [6,]  0.6013028  0.49334517  0.26703463 -0.30122519  0.6133917  1.0000000
#>  [7,]  0.4948550  0.33150343  0.45375974 -0.12606958  0.4879116  0.6706751
#>  [8,] -0.2494128 -0.35319949  0.24860459  0.33697653 -0.3118619 -0.4323780
#>  [9,]  0.5438826  0.30522926  0.60347329 -0.04629193  0.4993949  0.7042826
#>             [,7]       [,8]        [,9]
#>  [1,]  0.4948550 -0.2494128  0.54388257
#>  [2,]  0.3315034 -0.3531995  0.30522926
#>  [3,]  0.4537597  0.2486046  0.60347329
#>  [4,] -0.1260696  0.3369765 -0.04629193
#>  [5,]  0.4879116 -0.3118619  0.49939490
#>  [6,]  0.6706751 -0.4323780  0.70428262
#>  [7,]  1.0000000 -0.2086019  0.66509049
#>  [8,] -0.2086019  1.0000000 -0.15450840
#>  [9,]  0.6650905 -0.1545084  1.00000000