Calculate the residual correlation matrix from a latent variable model (LVM)

This function use coefficients \((\lambda_{jl} with j=1,\dots,n_species and l=1,\dots,n_latent)\), corresponding to latent variables fitted using jSDM package, to calculate the variance-covariance matrix which controls correlation between species.

get_residual_cor(mod)

Arguments

mod
An object of class "jSDM"

Value

results A list including :

cov.mean	Average over the MCMC samples of the variance-covariance matrix

cov.median	Median over the MCMC samples of the variance-covariance matrix

cor.mean	Average over the MCMC samples of the residual correlation matrix

cor.median	Median over the MCMC samples of the residual correlation matrix

Details

After fitting the jSDM with latent variables, the fullspecies residual correlation matrix : \(R=(R_ij) avec i=1,\ldots, nspecies et j=1,\ldots, nspecies\) can be derived from the covariance in the latent variables such as :

\(\Sigma_{ij}\)	\(= \lambda_i .\lambda_j' + 1\)	if i=j

\(= \lambda_i .\lambda_j'\)	else,

then we compute correlations from covariances : $$R_{ij} = \frac{\Sigma_{ij}}{\sqrt{\Sigma_ii\Sigma _jj}}$$.

References

Hui FKC (2016). “boral: Bayesian Ordination and Regression Analysis of Multivariate Abundance Data in R.” Methods in Ecology and Evolution, 7, 744–750.

Ovaskainen et al. (2016). Using latent variable models to identify large networks of species-to-species associations at different spatial scales. Methods in Ecology and Evolution, 7, 549-555.

Pollock et al. (2014). Understanding co-occurrence by modelling species simultaneously with a Joint Species Distribution Model (JSDM). Methods in Ecology and Evolution, 5, 397-406.

See also

Author

Ghislain Vieilledent <ghislain.vieilledent@cirad.fr>

Jeanne Clément <jeanne.clement16@laposte.net>

Examples

library(jSDM)
# frogs data
 data(frogs, package="jSDM")
 # Arranging data
 PA_frogs <- frogs[,4:12]
 # Normalized continuous variables
 Env_frogs <- cbind(scale(frogs[,1]),frogs[,2],scale(frogs[,3]))
 colnames(Env_frogs) <- colnames(frogs[,1:3])
 Env_frogs <- as.data.frame(Env_frogs)
 # Parameter inference
# Increase the number of iterations to reach MCMC convergence
mod <- jSDM_binomial_probit_block(# Response variable
                                  presence_site_sp=PA_frogs,
                                  # Explanatory variables
                                  site_suitability = ~.,
                                  site_data = Env_frogs,
                                  n_latent=2,
                                  site_effect="random",
                                  # Chains
                                  burnin=100,
                                  mcmc=100,
                                  thin=1,
                                  # Starting values
                                  alpha_start=0,
                                  beta_start=0,
                                  lambda_start=0,
                                  W_start=0,
                                  V_alpha=1,
                                  # Priors
                                  shape=0.5, rate=0.0005,
                                  mu_beta=0, V_beta=10,
                                  mu_lambda=0, V_lambda=10,
                                  # Various
                                  seed=1234, verbose=1)

#> 
#> Running the Gibbs sampler. It may be long, please keep cool :)
#> 
#> **********:10.0% 
#> **********:20.0% 
#> **********:30.0% 
#> **********:40.0% 
#> **********:50.0% 
#> **********:60.0% 
#> **********:70.0% 
#> **********:80.0% 
#> **********:90.0% 
#> **********:100.0% 

# Calcul of residual correlation between species 
 result <- get_residual_cor(mod)
 result$cov.mean

#>             [,1]       [,2]        [,3]       [,4]        [,5]      [,6]
#>  [1,]  2.3227534  0.5763644  0.63152910 -1.0446206  1.85106641  3.505928
#>  [2,]  0.5763644  1.3698234  0.01813650 -0.4982192  0.98905601  1.513773
#>  [3,]  0.6315291  0.0181365  3.24808746 -0.2519482 -0.03270329  1.853771
#>  [4,] -1.0446206 -0.4982192 -0.25194818  2.1304976 -1.66622924 -2.985769
#>  [5,]  1.8510664  0.9890560 -0.03270329 -1.6662292  4.71618879  4.892031
#>  [6,]  3.5059280  1.5137726  1.85377059 -2.9857689  4.89203131 11.093739
#>  [7,]  0.8671504  0.4337452  0.04193335 -0.7734189  1.37784092  2.539212
#>  [8,] -0.5836825 -0.2614004 -0.20439991  0.4887633 -0.87698217 -1.600147
#>  [9,]  2.3342480  0.8103615  2.50510416 -1.7266661  2.43046594  6.763256
#>              [,7]       [,8]       [,9]
#>  [1,]  0.86715041 -0.5836825  2.3342480
#>  [2,]  0.43374517 -0.2614004  0.8103615
#>  [3,]  0.04193335 -0.2043999  2.5051042
#>  [4,] -0.77341887  0.4887633 -1.7266661
#>  [5,]  1.37784092 -0.8769822  2.4304659
#>  [6,]  2.53921169 -1.6001467  6.7632555
#>  [7,]  2.00505477 -0.4195782  1.5151909
#>  [8,] -0.41957822  1.3445458 -1.0427275
#>  [9,]  1.51519090 -1.0427275  6.7626158

 result$cor.mean

#>             [,1]         [,2]         [,3]        [,4]        [,5]       [,6]
#>  [1,]  1.0000000  0.295041720  0.228518681 -0.43495841  0.54536329  0.6656969
#>  [2,]  0.2950417  1.000000000  0.008406505 -0.27530750  0.37832456  0.3752959
#>  [3,]  0.2285187  0.008406505  1.000000000 -0.08275954 -0.02370192  0.3041764
#>  [4,] -0.4349584 -0.275307504 -0.082759539  1.00000000 -0.50431077 -0.5649497
#>  [5,]  0.5453633  0.378324560 -0.023701919 -0.50431077  1.00000000  0.6977163
#>  [6,]  0.6656969  0.375295932  0.304176384 -0.56494971  0.69771632  1.0000000
#>  [7,]  0.3889339  0.247200395  0.029087938 -0.35658894  0.43815181  0.5271702
#>  [8,] -0.3097569 -0.183003771 -0.100185424  0.27388076 -0.33803375 -0.4065169
#>  [9,]  0.5678141  0.254122512  0.553098718 -0.41535022  0.45140844  0.7591981
#>              [,7]       [,8]       [,9]
#>  [1,]  0.38893394 -0.3097569  0.5678141
#>  [2,]  0.24720039 -0.1830038  0.2541225
#>  [3,]  0.02908794 -0.1001854  0.5530987
#>  [4,] -0.35658894  0.2738808 -0.4153502
#>  [5,]  0.43815181 -0.3380338  0.4514084
#>  [6,]  0.52717016 -0.4065169  0.7591981
#>  [7,]  1.00000000 -0.2453937  0.3617679
#>  [8,] -0.24539370  1.0000000 -0.3280921
#>  [9,]  0.36176793 -0.3280921  1.0000000