Calculate the residual correlation matrix from a latent variable model (LVM)

This function use coefficients \((\lambda_{jl} with j=1,\dots,n_species and l=1,\dots,n_latent)\), corresponding to latent variables fitted using jSDM package, to calculate the variance-covariance matrix which controls correlation between species.

get_residual_cor(mod)

Arguments

mod

An object of class "jSDM"

Value

results A list including :

cov.mean

Average over the MCMC samples of the variance-covariance matrix.

cov.median

Median over the MCMC samples of the variance-covariance matrix.

cor.mean

Average over the MCMC samples of the residual correlation matrix.

cor.median

Median over the MCMC samples of the residual correlation matrix.

Details

After fitting the jSDM with latent variables, the fullspecies residual correlation matrix : \(R=(R_ij) avec i=1,\ldots, nspecies et j=1,\ldots, nspecies\) can be derived from the covariance in the latent variables such as :

\(\Sigma_{ij}\)	\(= \lambda_i .\lambda_j' + 1\)	if i=j

\(= \lambda_i .\lambda_j'\)	else,

then we compute correlations from covariances : $$R_{ij} = \frac{\Sigma_{ij}}{\sqrt{\Sigma_ii\Sigma _jj}}$$.

References

Hui FKC (2016). “boral: Bayesian Ordination and Regression Analysis of Multivariate Abundance Data in R.” Methods in Ecology and Evolution, 7, 744–750.

Ovaskainen et al. (2016). Using latent variable models to identify large networks of species-to-species associations at different spatial scales. Methods in Ecology and Evolution, 7, 549-555.

Pollock et al. (2014). Understanding co-occurrence by modelling species simultaneously with a Joint Species Distribution Model (JSDM). Methods in Ecology and Evolution, 5, 397-406.

See also

Author

Ghislain Vieilledent <ghislain.vieilledent@cirad.fr>

Jeanne Clément <jeanne.clement16@laposte.net>

Examples

library(jSDM)
# frogs data
 data(frogs, package="jSDM")
 # Arranging data
 PA_frogs <- frogs[,4:12]
 # Normalized continuous variables
 Env_frogs <- cbind(scale(frogs[,1]),frogs[,2],scale(frogs[,3]))
 colnames(Env_frogs) <- colnames(frogs[,1:3])
 Env_frogs <- as.data.frame(Env_frogs)
 # Parameter inference
# Increase the number of iterations to reach MCMC convergence
 mod <- jSDM_binomial_probit(# Response variable
                             presence_data=PA_frogs,
                             # Explanatory variables
                             site_formula = ~.,
                             site_data = Env_frogs,
                             n_latent=2,
                             site_effect="random",
                             # Chains
                             burnin=100,
                             mcmc=100,
                             thin=1,
                             # Starting values
                             alpha_start=0,
                             beta_start=0,
                             lambda_start=0,
                             W_start=0,
                             V_alpha=1,
                             # Priors
                             shape=0.5, rate=0.0005,
                             mu_beta=0, V_beta=10,
                             mu_lambda=0, V_lambda=10,
                             # Various
                             seed=1234, verbose=1)
#> 
#> Running the Gibbs sampler. It may be long, please keep cool :)
#> 
#> **********:10.0% 
#> **********:20.0% 
#> **********:30.0% 
#> **********:40.0% 
#> **********:50.0% 
#> **********:60.0% 
#> **********:70.0% 
#> **********:80.0% 
#> **********:90.0% 
#> **********:100.0% 
# Calcul of residual correlation between species 
 result <- get_residual_cor(mod)
 result$cov.mean
#>             [,1]       [,2]       [,3]       [,4]       [,5]       [,6]
#>  [1,] 1.88651417 0.42707430  3.0891104  0.7910278  0.4422630  0.5098426
#>  [2,] 0.42707430 1.35579884  1.8758170  0.5291296  0.2111268  0.1872100
#>  [3,] 3.08911038 1.87581698 14.0941006  2.3991099  2.0502945  2.7399401
#>  [4,] 0.79102778 0.52912959  2.3991099  3.3079084 -0.3411853 -1.2460042
#>  [5,] 0.44226300 0.21112680  2.0502945 -0.3411853  1.8717346  1.2523301
#>  [6,] 0.50984264 0.18720997  2.7399401 -1.2460042  1.2523301  3.6224175
#>  [7,] 0.09649469 0.05421798  0.7904238 -0.4955056  0.4641193  0.8933733
#>  [8,] 0.30831288 0.21138642  1.2583301  0.6515816 -0.2543942 -0.1956588
#>  [9,] 1.31041778 0.70726630  6.4727710 -0.7900288  1.8072864  3.5789830
#>              [,7]       [,8]       [,9]
#>  [1,]  0.09649469  0.3083129  1.3104178
#>  [2,]  0.05421798  0.2113864  0.7072663
#>  [3,]  0.79042379  1.2583301  6.4727710
#>  [4,] -0.49550561  0.6515816 -0.7900288
#>  [5,]  0.46411927 -0.2543942  1.8072864
#>  [6,]  0.89337333 -0.1956588  3.5789830
#>  [7,]  1.50972108 -0.1838248  1.2493127
#>  [8,] -0.18382477  1.9813527  0.4221238
#>  [9,]  1.24931267  0.4221238  7.0243817
 result$cor.mean
#>             [,1]       [,2]      [,3]       [,4]        [,5]        [,6]
#>  [1,] 1.00000000 0.26348417 0.5845973  0.2870249  0.20842615  0.19393369
#>  [2,] 0.26348417 1.00000000 0.4152815  0.2495831  0.11715353  0.07986601
#>  [3,] 0.58459727 0.41528149 1.0000000  0.3492276  0.37299634  0.39261149
#>  [4,] 0.28702486 0.24958312 0.3492276  1.0000000 -0.14151734 -0.35530450
#>  [5,] 0.20842615 0.11715353 0.3729963 -0.1415173  1.00000000  0.46150151
#>  [6,] 0.19393369 0.07986601 0.3926115 -0.3553045  0.46150151  1.00000000
#>  [7,] 0.05800416 0.02743642 0.1522625 -0.2218658  0.24660404  0.36176260
#>  [8,] 0.16185432 0.12725605 0.2468859  0.2115327 -0.07976029 -0.04791613
#>  [9,] 0.37058190 0.21853109 0.6585472 -0.1665244  0.50363601  0.71686564
#>              [,7]        [,8]        [,9]
#>  [1,]  0.05800416  0.16185432  0.37058190
#>  [2,]  0.02743642  0.12725605  0.21853109
#>  [3,]  0.15226249  0.24688586  0.65854720
#>  [4,] -0.22186577  0.21153268 -0.16652443
#>  [5,]  0.24660404 -0.07976029  0.50363601
#>  [6,]  0.36176260 -0.04791613  0.71686564
#>  [7,]  1.00000000 -0.09164662  0.35491504
#>  [8,] -0.09164662  1.00000000  0.07761542
#>  [9,]  0.35491504  0.07761542  1.00000000